swap out wasmer packages
