+++ lib/runtime/oldWasmer/config.go
@@ -1,7 +1,7 @@
-// Copyright 2023 ChainSafe Systems (ON)
+// Copyright 2022 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
-package newWasmer
+package oldWasmer
 
 import (
 	"testing"
@@ -12,7 +12,8 @@ import (
 	"github.com/ChainSafe/gossamer/lib/runtime"
 )
 
-// Config is the configuration used to create a Wasmer runtime instance.
+// Config is the configuration used to create a
+// Wasmer runtime instance.
 type Config struct {
 	Storage     Storage
 	Keystore    *keystore.GlobalKeystore
