+++ lib/runtime/oldWasmer/exports_test.go
@@ -1,11 +1,13 @@
-// Copyright 2023 ChainSafe Systems (ON)
+// Copyright 2021 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
-package newWasmer
+package oldWasmer
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
+	"math/big"
 	"os"
 	"testing"
 
@@ -15,11 +17,13 @@ import (
 	"github.com/ChainSafe/gossamer/lib/crypto/ed25519"
 	"github.com/ChainSafe/gossamer/lib/genesis"
 	"github.com/ChainSafe/gossamer/lib/runtime"
+	"github.com/ChainSafe/gossamer/lib/runtime/oldWasmer/testdata"
 	"github.com/ChainSafe/gossamer/lib/runtime/storage"
-	"github.com/ChainSafe/gossamer/lib/runtime/wasmer/testdata"
 	"github.com/ChainSafe/gossamer/lib/trie"
 	"github.com/ChainSafe/gossamer/lib/utils"
 	"github.com/ChainSafe/gossamer/pkg/scale"
+	"github.com/centrifuge/go-substrate-rpc-client/v4/signature"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -29,7 +33,7 @@ import (
 var testKeyOwnershipProof types.OpaqueKeyOwnershipProof = types.OpaqueKeyOwnershipProof([]byte{64, 138, 252, 29, 127, 102, 189, 129, 207, 47, 157, 60, 17, 138, 194, 121, 139, 92, 176, 175, 224, 16, 185, 93, 175, 251, 224, 81, 209, 61, 0, 71}) //nolint:lll
 
 func Test_Instance_Version(t *testing.T) {
-	//t.Parallel()
+	t.Parallel()
 
 	type instanceVersioner interface {
 		Version() (version runtime.Version)
@@ -278,7 +282,7 @@ func Test_Instance_Version(t *testing.T) {
 	for name, testCase := range testCases {
 		testCase := testCase
 		t.Run(name, func(t *testing.T) {
-			//t.Parallel()
+			t.Parallel()
 
 			instance := testCase.instanceBuilder(t)
 			version := instance.Version()
@@ -297,69 +301,68 @@ func balanceKey(t *testing.T, pub []byte) []byte {
 	return append(append(append(h0, h1...), h2...), pub...)
 }
 
-// TODO fix
-//func TestNodeRuntime_ValidateTransaction(t *testing.T) {
-//	genesisPath := utils.GetWestendDevRawGenesisPath(t)
-//	gen := genesisFromRawJSON(t, genesisPath)
-//	genTrie, err := NewTrieFromGenesis(gen)
-//	require.NoError(t, err)
-//
-//	// set state to genesis state
-//	genState := storage.NewTrieState(&genTrie)
-//
-//	cfg := Config{
-//		Storage: genState,
-//		LogLvl:  log.Critical,
-//	}
-//
-//	nodeStorage := runtime.NodeStorage{}
-//	nodeStorage.BaseDB = runtime.NewInMemoryDB(t)
-//	cfg.NodeStorage = nodeStorage
-//
-//	rt, err := NewRuntimeFromGenesis(cfg)
-//	require.NoError(t, err)
-//
-//	alicePub := common.MustHexToBytes("0xd43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d")
-//	aliceBalanceKey := balanceKey(t, alicePub)
-//
-//	accInfo := types.AccountInfo{
-//		Nonce: 0,
-//		Data: types.AccountData{
-//			Free:       scale.MustNewUint128(big.NewInt(1152921504606846976)),
-//			Reserved:   scale.MustNewUint128(big.NewInt(0)),
-//			MiscFrozen: scale.MustNewUint128(big.NewInt(0)),
-//			FreeFrozen: scale.MustNewUint128(big.NewInt(0)),
-//		},
-//	}
-//
-//	encBal, err := scale.Marshal(accInfo)
-//	require.NoError(t, err)
-//
-//	rt.ctx.Storage.Put(aliceBalanceKey, encBal)
-//	// this key is System.UpgradedToDualRefCount -> set to true since all accounts have been upgraded to v0.9 format
-//	rt.ctx.Storage.Put(common.UpgradedToDualRefKey, []byte{1})
-//
-//	genesisHeader := &types.Header{
-//		Number:    0,
-//		StateRoot: genTrie.MustHash(),
-//	}
-//
-//	extHex := runtime.NewTestExtrinsic(t, rt, genesisHeader.Hash(), genesisHeader.Hash(),
-//		0, signature.TestKeyringPairAlice, "System.remark", []byte{0xab, 0xcd})
-//
-//	genesisHashBytes := genesisHeader.Hash().ToBytes()
-//
-//	validateTransactionArguments := [][]byte{
-//		{byte(types.TxnExternal)},
-//		common.MustHexToBytes(extHex),
-//		genesisHashBytes}
-//
-//	extrinsicsBytes := bytes.Join(validateTransactionArguments, nil)
-//
-//	runtime.InitializeRuntimeToTest(t, rt, genesisHeader)
-//	_, err = rt.ValidateTransaction(extrinsicsBytes)
-//	require.NoError(t, err)
-//}
+func TestNodeRuntime_ValidateTransaction(t *testing.T) {
+	genesisPath := utils.GetWestendDevRawGenesisPath(t)
+	gen := genesisFromRawJSON(t, genesisPath)
+	genTrie, err := NewTrieFromGenesis(gen)
+	require.NoError(t, err)
+
+	// set state to genesis state
+	genState := storage.NewTrieState(&genTrie)
+
+	cfg := Config{
+		Storage: genState,
+		LogLvl:  log.Critical,
+	}
+
+	nodeStorage := runtime.NodeStorage{}
+	nodeStorage.BaseDB = runtime.NewInMemoryDB(t)
+	cfg.NodeStorage = nodeStorage
+
+	rt, err := NewRuntimeFromGenesis(cfg)
+	require.NoError(t, err)
+
+	alicePub := common.MustHexToBytes("0xd43593c715fdd31c61141abd04a99fd6822c8558854ccde39a5684e7a56da27d")
+	aliceBalanceKey := balanceKey(t, alicePub)
+
+	accInfo := types.AccountInfo{
+		Nonce: 0,
+		Data: types.AccountData{
+			Free:       scale.MustNewUint128(big.NewInt(1152921504606846976)),
+			Reserved:   scale.MustNewUint128(big.NewInt(0)),
+			MiscFrozen: scale.MustNewUint128(big.NewInt(0)),
+			FreeFrozen: scale.MustNewUint128(big.NewInt(0)),
+		},
+	}
+
+	encBal, err := scale.Marshal(accInfo)
+	require.NoError(t, err)
+
+	rt.ctx.Storage.Put(aliceBalanceKey, encBal)
+	// this key is System.UpgradedToDualRefCount -> set to true since all accounts have been upgraded to v0.9 format
+	rt.ctx.Storage.Put(common.UpgradedToDualRefKey, []byte{1})
+
+	genesisHeader := &types.Header{
+		Number:    0,
+		StateRoot: genTrie.MustHash(),
+	}
+
+	extHex := runtime.NewTestExtrinsic(t, rt, genesisHeader.Hash(), genesisHeader.Hash(),
+		0, signature.TestKeyringPairAlice, "System.remark", []byte{0xab, 0xcd})
+
+	genesisHashBytes := genesisHeader.Hash().ToBytes()
+
+	validateTransactionArguments := [][]byte{
+		{byte(types.TxnExternal)},
+		common.MustHexToBytes(extHex),
+		genesisHashBytes}
+
+	extrinsicsBytes := bytes.Join(validateTransactionArguments, nil)
+
+	runtime.InitializeRuntimeToTest(t, rt, genesisHeader)
+	_, err = rt.ValidateTransaction(extrinsicsBytes)
+	require.NoError(t, err)
+}
 
 func TestInstance_GrandpaAuthorities_NodeRuntime(t *testing.T) {
 	tt := trie.NewEmptyTrie()
@@ -418,7 +421,7 @@ func TestInstance_GrandpaAuthorities_PolkadotRuntime(t *testing.T) {
 }
 
 func TestInstance_BabeGenerateKeyOwnershipProof(t *testing.T) {
-	//t.Parallel()
+	t.Parallel()
 
 	testCases := []struct {
 		name          string
@@ -436,7 +439,7 @@ func TestInstance_BabeGenerateKeyOwnershipProof(t *testing.T) {
 	for _, testCase := range testCases {
 		testCase := testCase
 		t.Run(testCase.name, func(t *testing.T) {
-			//t.Parallel()
+			t.Parallel()
 
 			tt := trie.NewEmptyTrie()
 
@@ -468,7 +471,7 @@ func TestInstance_BabeGenerateKeyOwnershipProof(t *testing.T) {
 }
 
 func TestInstance_BabeSubmitReportEquivocationUnsignedExtrinsic(t *testing.T) {
-	//t.Parallel()
+	t.Parallel()
 
 	testCases := []struct {
 		name          string
@@ -486,7 +489,7 @@ func TestInstance_BabeSubmitReportEquivocationUnsignedExtrinsic(t *testing.T) {
 	for _, testCase := range testCases {
 		testCase := testCase
 		t.Run(testCase.name, func(t *testing.T) {
-			//t.Parallel()
+			t.Parallel()
 
 			tt := trie.NewEmptyTrie()
 			rt := NewTestInstanceWithTrie(t, testCase.targetRuntime, tt)
@@ -605,77 +608,74 @@ func TestInstance_InitializeBlock_PolkadotRuntime(t *testing.T) {
 	require.NoError(t, err)
 }
 
-// TODO fix
-//func TestInstance_ExecuteBlock_WestendRuntime(t *testing.T) {
-//	instance := NewTestInstance(t, runtime.WESTEND_RUNTIME_v0929)
-//	block := runtime.InitializeRuntimeToTest(t, instance, &types.Header{})
-//
-//	// reset state back to parent state before executing
-//	parentState := storage.NewTrieState(nil)
-//	instance.SetContextStorage(parentState)
-//
-//	_, err := instance.ExecuteBlock(block)
-//	require.NoError(t, err)
-//}
+func TestInstance_ExecuteBlock_WestendRuntime(t *testing.T) {
+	instance := NewTestInstance(t, runtime.WESTEND_RUNTIME_v0929)
+	block := runtime.InitializeRuntimeToTest(t, instance, &types.Header{})
 
-// TODO fix
-//func TestInstance_ApplyExtrinsic_WestendRuntime(t *testing.T) {
-//	genesisPath := utils.GetWestendDevRawGenesisPath(t)
-//	gen := genesisFromRawJSON(t, genesisPath)
-//	genTrie, err := NewTrieFromGenesis(gen)
-//	require.NoError(t, err)
-//
-//	// set state to genesis state
-//	genState := storage.NewTrieState(&genTrie)
-//
-//	cfg := Config{
-//		Storage: genState,
-//		LogLvl:  log.Critical,
-//	}
-//
-//	instance, err := NewRuntimeFromGenesis(cfg)
-//	require.NoError(t, err)
-//
-//	// reset state back to parent state before executing
-//	parentState := storage.NewTrieState(&genTrie)
-//	instance.SetContextStorage(parentState)
-//
-//	genesisHeader := &types.Header{
-//		Number:    0,
-//		StateRoot: genTrie.MustHash(),
-//	}
-//	header := &types.Header{
-//		ParentHash: genesisHeader.Hash(),
-//		Number:     1,
-//		Digest:     types.NewDigest(),
-//	}
-//
-//	err = instance.InitializeBlock(header)
-//	require.NoError(t, err)
-//
-//	extHex := runtime.NewTestExtrinsic(t, instance, genesisHeader.Hash(), genesisHeader.Hash(),
-//		0, signature.TestKeyringPairAlice, "System.remark", []byte{0xab, 0xcd})
-//
-//	res, err := instance.ApplyExtrinsic(common.MustHexToBytes(extHex))
-//	require.NoError(t, err)
-//	require.Equal(t, []byte{0, 0}, res)
-//}
+	// reset state back to parent state before executing
+	parentState := storage.NewTrieState(nil)
+	instance.SetContextStorage(parentState)
 
-// TODO fix
-//func TestInstance_ExecuteBlock_PolkadotRuntime(t *testing.T) {
-//	DefaultTestLogLvl = 0
-//
-//	instance := NewTestInstance(t, runtime.POLKADOT_RUNTIME)
-//
-//	block := runtime.InitializeRuntimeToTest(t, instance, &types.Header{})
-//
-//	// reset state back to parent state before executing
-//	parentState := storage.NewTrieState(nil)
-//	instance.SetContextStorage(parentState)
-//
-//	_, err := instance.ExecuteBlock(block)
-//	require.NoError(t, err)
-//}
+	_, err := instance.ExecuteBlock(block)
+	require.NoError(t, err)
+}
+
+func TestInstance_ApplyExtrinsic_WestendRuntime(t *testing.T) {
+	genesisPath := utils.GetWestendDevRawGenesisPath(t)
+	gen := genesisFromRawJSON(t, genesisPath)
+	genTrie, err := NewTrieFromGenesis(gen)
+	require.NoError(t, err)
+
+	// set state to genesis state
+	genState := storage.NewTrieState(&genTrie)
+
+	cfg := Config{
+		Storage: genState,
+		LogLvl:  log.Critical,
+	}
+
+	instance, err := NewRuntimeFromGenesis(cfg)
+	require.NoError(t, err)
+
+	// reset state back to parent state before executing
+	parentState := storage.NewTrieState(&genTrie)
+	instance.SetContextStorage(parentState)
+
+	genesisHeader := &types.Header{
+		Number:    0,
+		StateRoot: genTrie.MustHash(),
+	}
+	header := &types.Header{
+		ParentHash: genesisHeader.Hash(),
+		Number:     1,
+		Digest:     types.NewDigest(),
+	}
+
+	err = instance.InitializeBlock(header)
+	require.NoError(t, err)
+
+	extHex := runtime.NewTestExtrinsic(t, instance, genesisHeader.Hash(), genesisHeader.Hash(),
+		0, signature.TestKeyringPairAlice, "System.remark", []byte{0xab, 0xcd})
+
+	res, err := instance.ApplyExtrinsic(common.MustHexToBytes(extHex))
+	require.NoError(t, err)
+	require.Equal(t, []byte{0, 0}, res)
+}
+
+func TestInstance_ExecuteBlock_PolkadotRuntime(t *testing.T) {
+	DefaultTestLogLvl = 0
+
+	instance := NewTestInstance(t, runtime.POLKADOT_RUNTIME)
+
+	block := runtime.InitializeRuntimeToTest(t, instance, &types.Header{})
+
+	// reset state back to parent state before executing
+	parentState := storage.NewTrieState(nil)
+	instance.SetContextStorage(parentState)
+
+	_, err := instance.ExecuteBlock(block)
+	require.NoError(t, err)
+}
 
 func TestInstance_ExecuteBlock_PolkadotRuntime_PolkadotBlock1(t *testing.T) {
 	genesisPath := utils.GetPolkadotGenesisPath(t)
@@ -1095,13 +1095,15 @@ func TestInstance_PaymentQueryInfo(t *testing.T) {
 		},
 		{
 			// incomplete extrinsic
-			ext:        "0x4ccde39a5684e7a56da23b22d4d9fbadb023baa19c56495432884d0640000000000000000000000000000000",
-			errMessage: "running runtime function: unreachable",
+			ext: "0x4ccde39a5684e7a56da23b22d4d9fbadb023baa19c56495432884d0640000000000000000000000000000000",
+			errMessage: "running runtime function: " +
+				"Failed to call the `TransactionPaymentApi_query_info` exported function.",
 		},
 		{
 			// incomplete extrinsic
-			extB:       nil,
-			errMessage: "running runtime function: unreachable",
+			extB: nil,
+			errMessage: "running runtime function: " +
+				"Failed to call the `TransactionPaymentApi_query_info` exported function.",
 		},
 	}
 
@@ -1155,7 +1157,7 @@ func newTrieFromPairs(t *testing.T, filename string) *trie.Trie {
 }
 
 func TestInstance_TransactionPaymentCallApi_QueryCallInfo(t *testing.T) {
-	//t.Parallel()
+	t.Parallel()
 	ins := NewTestInstance(t, runtime.WESTEND_RUNTIME_v0929)
 	tests := []struct {
 		callHex    string
@@ -1180,15 +1182,17 @@ func TestInstance_TransactionPaymentCallApi_QueryCallInfo(t *testing.T) {
 			// polkadot.js/api returns error:  RPC-CORE: call(method: Text, data: Bytes, at?: BlockHash):
 			// Bytes:: -32000: Client error: Execution failed: Execution aborted due to trap: wasm trap: wasm
 			//`unreachable` instruction executed
-			callHex:    "0x040001084564",
-			errMessage: "running runtime function: unreachable",
+			callHex: "0x040001084564",
+			errMessage: "running runtime function: " +
+				"Failed to call the `TransactionPaymentCallApi_query_call_info` exported function.",
 		},
 		{
 			// call without removing any bytes, polkadot.js/api v9.5.1: api.tx.system.remark("Ed test")
 			// polkadot.js/api returns error: Error: createType(Call):: findMetaCall: Unable to find Call with index
 			// [44, 4]/[44,4]
-			callHex:    "0x2c0400011c45642074657374",
-			errMessage: "running runtime function: unreachable",
+			callHex: "0x2c0400011c45642074657374",
+			errMessage: "running runtime function: " +
+				"Failed to call the `TransactionPaymentCallApi_query_call_info` exported function.",
 		},
 	}
 
@@ -1213,7 +1217,7 @@ func TestInstance_TransactionPaymentCallApi_QueryCallInfo(t *testing.T) {
 }
 
 func TestInstance_TransactionPaymentCallApi_QueryCallFeeDetails(t *testing.T) {
-	//t.Parallel()
+	t.Parallel()
 	ins := NewTestInstance(t, runtime.WESTEND_RUNTIME_v0929)
 	tests := []struct {
 		callHex    string
@@ -1244,15 +1248,17 @@ func TestInstance_TransactionPaymentCallApi_QueryCallFeeDetails(t *testing.T) {
 			// when calling polkadot node (v0.9.29) with polkadot.js/api the node returns error:  RPC-CORE: call(
 			// method: Text, data: Bytes, at?: BlockHash): Bytes:: -32000: Client error: Execution failed:
 			// Execution aborted due to trap: wasm trap: wasm `unreachable` instruction executed
-			callHex:    "0x040001084564",
-			errMessage: "running runtime function: unreachable",
+			callHex: "0x040001084564",
+			errMessage: "running runtime function: " +
+				"Failed to call the `TransactionPaymentCallApi_query_call_fee_details` exported function.",
 		},
 		{
 			// call without removing any bytes, polkadot.js/api v9.5.1: api.tx.system.remark("Ed test")
 			// when calling polkadot (v0.9.29) with polkadot.js/api the node returns error: Error: createType(
 			//Call):: findMetaCall: Unable to find Call with index [44, 4]/[44,4]
-			callHex:    "0x18040001084564",
-			errMessage: "running runtime function: unreachable",
+			callHex: "0x18040001084564",
+			errMessage: "running runtime function: " +
+				"Failed to call the `TransactionPaymentCallApi_query_call_fee_details` exported function.",
 		},
 	}
 
@@ -1274,7 +1280,7 @@ func TestInstance_TransactionPaymentCallApi_QueryCallFeeDetails(t *testing.T) {
 }
 
 func TestInstance_GrandpaGenerateKeyOwnershipProof(t *testing.T) {
-	//t.Parallel()
+	t.Parallel()
 	instance := NewTestInstance(t, runtime.WESTEND_RUNTIME_v0929)
 	identity := common.MustHexToBytes("0x88dc3417d5058ec4b4503e0c12ea1a0a89be200fe98922423d4334014fa6b0ee")
 	identityPubKey, _ := ed25519.NewPublicKey(identity)
@@ -1287,7 +1293,7 @@ func TestInstance_GrandpaGenerateKeyOwnershipProof(t *testing.T) {
 }
 
 func TestInstance_GrandpaSubmitReportEquivocationUnsignedExtrinsic(t *testing.T) {
-	//t.Parallel()
+	t.Parallel()
 	identity := common.MustHexToBytes("0x88dc3417d5058ec4b4503e0c12ea1a0a89be200fe98922423d4334014fa6b0ee")
 	identityPubKey, _ := ed25519.NewPublicKey(identity)
 	runtime := NewTestInstance(t, runtime.WESTEND_RUNTIME_v0929)
