+++ lib/runtime/wasmer/config_test.go
@@ -1,4 +1,4 @@
-// Copyright 2022 ChainSafe Systems (ON)
+// Copyright 2023 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
 package wasmer
