+++ lib/runtime/wasmer/imports.go
@@ -1,114 +1,11 @@
-// Copyright 2021 ChainSafe Systems (ON)
+// Copyright 2023 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
 package wasmer
 
-// #include <stdlib.h>
-//
-// extern void ext_logging_log_version_1(void *context, int32_t level, int64_t target, int64_t msg);
-// extern int32_t ext_logging_max_level_version_1(void *context);
-//
-// extern void ext_sandbox_instance_teardown_version_1(void *context, int32_t a);
-// extern int32_t ext_sandbox_instantiate_version_1(void *context, int32_t a, int64_t b, int64_t c, int32_t d);
-// extern int32_t ext_sandbox_invoke_version_1(void *context, int32_t a, int64_t b, int64_t c, int32_t d, int32_t e, int32_t f);
-// extern int32_t ext_sandbox_memory_get_version_1(void *context, int32_t a, int32_t b, int32_t c, int32_t d);
-// extern int32_t ext_sandbox_memory_new_version_1(void *context, int32_t a, int32_t b);
-// extern int32_t ext_sandbox_memory_set_version_1(void *context, int32_t a, int32_t b, int32_t c, int32_t d);
-// extern void ext_sandbox_memory_teardown_version_1(void *context, int32_t a);
-//
-// extern int32_t ext_crypto_ed25519_generate_version_1(void *context, int32_t a, int64_t b);
-// extern int64_t ext_crypto_ed25519_public_keys_version_1(void *context, int32_t a);
-// extern int64_t ext_crypto_ed25519_sign_version_1(void *context, int32_t a, int32_t b, int64_t c);
-// extern int32_t ext_crypto_ed25519_verify_version_1(void *context, int32_t a, int64_t b, int32_t c);
-// extern int32_t ext_crypto_finish_batch_verify_version_1(void *context);
-// extern int64_t ext_crypto_secp256k1_ecdsa_recover_version_1(void *context, int32_t a, int32_t b);
-// extern int64_t ext_crypto_secp256k1_ecdsa_recover_version_2(void *context, int32_t a, int32_t b);
-// extern int64_t ext_crypto_secp256k1_ecdsa_recover_compressed_version_1(void *context, int32_t a, int32_t b);
-// extern int64_t ext_crypto_secp256k1_ecdsa_recover_compressed_version_2(void *context, int32_t a, int32_t b);
-// extern int32_t ext_crypto_ecdsa_verify_version_2(void *context, int32_t a, int64_t b, int32_t c);
-// extern int32_t ext_crypto_sr25519_generate_version_1(void *context, int32_t a, int64_t b);
-// extern int64_t ext_crypto_sr25519_public_keys_version_1(void *context, int32_t a);
-// extern int64_t ext_crypto_sr25519_sign_version_1(void *context, int32_t a, int32_t b, int64_t c);
-// extern int32_t ext_crypto_sr25519_verify_version_1(void *context, int32_t a, int64_t b, int32_t c);
-// extern int32_t ext_crypto_sr25519_verify_version_2(void *context, int32_t a, int64_t b, int32_t c);
-// extern void ext_crypto_start_batch_verify_version_1(void *context);
-//
-// extern int32_t ext_trie_blake2_256_root_version_1(void *context, int64_t a);
-// extern int32_t ext_trie_blake2_256_ordered_root_version_1(void *context, int64_t a);
-// extern int32_t ext_trie_blake2_256_ordered_root_version_2(void *context, int64_t a, int32_t b);
-// extern int32_t ext_trie_blake2_256_verify_proof_version_1(void *context, int32_t a, int64_t b, int64_t c, int64_t d);
-//
-// extern int64_t ext_misc_runtime_version_version_1(void *context, int64_t a);
-// extern void ext_misc_print_hex_version_1(void *context, int64_t a);
-// extern void ext_misc_print_num_version_1(void *context, int64_t a);
-// extern void ext_misc_print_utf8_version_1(void *context, int64_t a);
-//
-// extern void ext_default_child_storage_clear_version_1(void *context, int64_t a, int64_t b);
-// extern int64_t ext_default_child_storage_get_version_1(void *context, int64_t a, int64_t b);
-// extern int64_t ext_default_child_storage_next_key_version_1(void *context, int64_t a, int64_t b);
-// extern int64_t ext_default_child_storage_read_version_1(void *context, int64_t a, int64_t b, int64_t c, int32_t d);
-// extern int64_t ext_default_child_storage_root_version_1(void *context, int64_t a);
-// extern void ext_default_child_storage_set_version_1(void *context, int64_t a, int64_t b, int64_t c);
-// extern void ext_default_child_storage_storage_kill_version_1(void *context, int64_t a);
-// extern int32_t ext_default_child_storage_storage_kill_version_2(void *context, int64_t a, int64_t b);
-// extern int64_t ext_default_child_storage_storage_kill_version_3(void *context, int64_t a, int64_t b);
-// extern void ext_default_child_storage_clear_prefix_version_1(void *context, int64_t a, int64_t b);
-// extern int32_t ext_default_child_storage_exists_version_1(void *context, int64_t a, int64_t b);
-//
-// extern void ext_allocator_free_version_1(void *context, int32_t a);
-// extern int32_t ext_allocator_malloc_version_1(void *context, int32_t a);
-//
-// extern int32_t ext_hashing_blake2_128_version_1(void *context, int64_t a);
-// extern int32_t ext_hashing_blake2_256_version_1(void *context, int64_t a);
-// extern int32_t ext_hashing_keccak_256_version_1(void *context, int64_t a);
-// extern int32_t ext_hashing_sha2_256_version_1(void *context, int64_t a);
-// extern int32_t ext_hashing_twox_256_version_1(void *context, int64_t a);
-// extern int32_t ext_hashing_twox_128_version_1(void *context, int64_t a);
-// extern int32_t ext_hashing_twox_64_version_1(void *context, int64_t a);
-//
-// extern void ext_offchain_index_set_version_1(void *context, int64_t a, int64_t b);
-// extern int32_t ext_offchain_is_validator_version_1(void *context);
-// extern void ext_offchain_local_storage_clear_version_1(void *context, int32_t a, int64_t b);
-// extern int32_t ext_offchain_local_storage_compare_and_set_version_1(void *context, int32_t a, int64_t b, int64_t c, int64_t d);
-// extern int64_t ext_offchain_local_storage_get_version_1(void *context, int32_t a, int64_t b);
-// extern void ext_offchain_local_storage_set_version_1(void *context, int32_t a, int64_t b, int64_t c);
-// extern int64_t ext_offchain_network_state_version_1(void *context);
-// extern int32_t ext_offchain_random_seed_version_1(void *context);
-// extern int64_t ext_offchain_submit_transaction_version_1(void *context, int64_t a);
-// extern int64_t ext_offchain_timestamp_version_1(void *context);
-// extern void ext_offchain_sleep_until_version_1(void *context, int64_t a);
-// extern int64_t ext_offchain_http_request_start_version_1(void *context, int64_t a, int64_t b, int64_t c);
-// extern int64_t ext_offchain_http_request_add_header_version_1(void *context, int32_t a, int64_t k, int64_t v);
-//
-// extern void ext_storage_append_version_1(void *context, int64_t a, int64_t b);
-// extern int64_t ext_storage_changes_root_version_1(void *context, int64_t a);
-// extern void ext_storage_clear_version_1(void *context, int64_t a);
-// extern void ext_storage_clear_prefix_version_1(void *context, int64_t a);
-// extern int64_t ext_storage_clear_prefix_version_2(void *context, int64_t a, int64_t b);
-// extern void ext_storage_commit_transaction_version_1(void *context);
-// extern int32_t ext_storage_exists_version_1(void *context, int64_t a);
-// extern int64_t ext_storage_get_version_1(void *context, int64_t a);
-// extern int64_t ext_storage_next_key_version_1(void *context, int64_t a);
-// extern int64_t ext_storage_read_version_1(void *context, int64_t a, int64_t b, int32_t c);
-// extern void ext_storage_rollback_transaction_version_1(void *context);
-// extern int64_t ext_storage_root_version_1(void *context);
-// extern int64_t ext_storage_root_version_2(void *context, int32_t a);
-// extern void ext_storage_set_version_1(void *context, int64_t a, int64_t b);
-// extern void ext_storage_start_transaction_version_1(void *context);
-//
-// extern void ext_transaction_index_index_version_1(void *context, int32_t a, int32_t b, int32_t c);
-// extern void ext_transaction_index_renew_version_1(void *context, int32_t a, int32_t b);
-import "C" //skipcq: SCC-compile
-
 import (
 	"encoding/binary"
 	"fmt"
-	"math/big"
-	"math/rand"
-	"reflect"
-	"time"
-	"unsafe"
-
 	"github.com/ChainSafe/gossamer/lib/common"
 	"github.com/ChainSafe/gossamer/lib/crypto"
 	"github.com/ChainSafe/gossamer/lib/crypto/ed25519"
@@ -119,8 +16,11 @@ import (
 	"github.com/ChainSafe/gossamer/lib/trie"
 	"github.com/ChainSafe/gossamer/lib/trie/proof"
 	"github.com/ChainSafe/gossamer/pkg/scale"
-
-	wasm "github.com/wasmerio/go-ext-wasm/wasmer"
+	"github.com/wasmerio/wasmer-go/wasmer"
+	"math/big"
+	"math/rand"
+	"reflect"
+	"time"
 )
 
 const (
@@ -128,12 +28,20 @@ const (
 )
 
 //export ext_logging_log_version_1
-func ext_logging_log_version_1(context unsafe.Pointer, level C.int32_t, targetData, msgData C.int64_t) {
+func ext_logging_log_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
+	ctx := env.(*runtime.Context)
 
-	target := string(asMemorySlice(instanceContext, targetData))
-	msg := string(asMemorySlice(instanceContext, msgData))
+	level, ok := args[0].Unwrap().(int32)
+	if !ok {
+		logger.Criticalf("[ext_logging_log_version_1]", "error", "addr cannot be converted to int32")
+	}
+
+	targetData := args[1].I64()
+	msgData := args[2].I64()
+
+	target := string(asMemorySlice(ctx, targetData))
+	msg := string(asMemorySlice(ctx, msgData))
 
 	switch int(level) {
 	case 0:
@@ -149,89 +57,96 @@ func ext_logging_log_version_1(context unsafe.Pointer, level C.int32_t, targetDa
 	default:
 		logger.Errorf("level=%d target=%s message=%s", int(level), target, msg)
 	}
+	return nil, nil
 }
 
 //export ext_logging_max_level_version_1
-func ext_logging_max_level_version_1(context unsafe.Pointer) C.int32_t {
+func ext_logging_max_level_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	return 4
+	return []wasmer.Value{wasmer.NewI32(4)}, nil
 }
 
 //export ext_transaction_index_index_version_1
-func ext_transaction_index_index_version_1(context unsafe.Pointer, a, b, c C.int32_t) {
+func ext_transaction_index_index_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 	logger.Warn("unimplemented")
+	return nil, nil
 }
 
 //export ext_transaction_index_renew_version_1
-func ext_transaction_index_renew_version_1(context unsafe.Pointer, a, b C.int32_t) {
+func ext_transaction_index_renew_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 	logger.Warn("unimplemented")
+	return nil, nil
 }
 
 //export ext_sandbox_instance_teardown_version_1
-func ext_sandbox_instance_teardown_version_1(context unsafe.Pointer, a C.int32_t) {
+func ext_sandbox_instance_teardown_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 	logger.Warn("unimplemented")
+	return nil, nil
 }
 
 //export ext_sandbox_instantiate_version_1
-func ext_sandbox_instantiate_version_1(context unsafe.Pointer, a C.int32_t, x, y C.int64_t, z C.int32_t) C.int32_t {
+func ext_sandbox_instantiate_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 	logger.Warn("unimplemented")
-	return 0
+	return []wasmer.Value{wasmer.NewI32(0)}, nil
 }
 
 //export ext_sandbox_invoke_version_1
-func ext_sandbox_invoke_version_1(context unsafe.Pointer, a C.int32_t, x, y C.int64_t, z, d, e C.int32_t) C.int32_t {
+func ext_sandbox_invoke_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 	logger.Warn("unimplemented")
-	return 0
+	return []wasmer.Value{wasmer.NewI32(0)}, nil
 }
 
 //export ext_sandbox_memory_get_version_1
-func ext_sandbox_memory_get_version_1(context unsafe.Pointer, a, z, d, e C.int32_t) C.int32_t {
+func ext_sandbox_memory_get_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 	logger.Warn("unimplemented")
-	return 0
+	return []wasmer.Value{wasmer.NewI32(0)}, nil
 }
 
 //export ext_sandbox_memory_new_version_1
-func ext_sandbox_memory_new_version_1(context unsafe.Pointer, a, z C.int32_t) C.int32_t {
+func ext_sandbox_memory_new_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 	logger.Warn("unimplemented")
-	return 0
+	return []wasmer.Value{wasmer.NewI32(0)}, nil
 }
 
 //export ext_sandbox_memory_set_version_1
-func ext_sandbox_memory_set_version_1(context unsafe.Pointer, a, z, d, e C.int32_t) C.int32_t {
+func ext_sandbox_memory_set_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 	logger.Warn("unimplemented")
-	return 0
+	return []wasmer.Value{wasmer.NewI32(0)}, nil
 }
 
 //export ext_sandbox_memory_teardown_version_1
-func ext_sandbox_memory_teardown_version_1(context unsafe.Pointer, a C.int32_t) {
+func ext_sandbox_memory_teardown_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 	logger.Warn("unimplemented")
+	return nil, nil
 }
 
 //export ext_crypto_ed25519_generate_version_1
-func ext_crypto_ed25519_generate_version_1(context unsafe.Pointer, keyTypeID C.int32_t, seedSpan C.int64_t) C.int32_t {
+func ext_crypto_ed25519_generate_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	memory := instanceContext.Memory().Data()
+	runtimeCtx := env.(*runtime.Context)
+	memory := runtimeCtx.Memory.Data()
+
+	keyTypeID := args[0].I32()
+	seedSpan := args[1].I64()
 
 	id := memory[keyTypeID : keyTypeID+4]
-	seedBytes := asMemorySlice(instanceContext, seedSpan)
+	seedBytes := asMemorySlice(runtimeCtx, seedSpan)
 
 	var seed *[]byte
 	err := scale.Unmarshal(seedBytes, &seed)
 	if err != nil {
 		logger.Warnf("cannot generate key: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	var kp KeyPair
@@ -244,54 +159,53 @@ func ext_crypto_ed25519_generate_version_1(context unsafe.Pointer, keyTypeID C.i
 
 	if err != nil {
 		logger.Warnf("cannot generate key: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	ks, err := runtimeCtx.Keystore.GetKeystore(id)
 	if err != nil {
 		logger.Warnf("error for id 0x%x: %s", id, err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	err = ks.Insert(kp)
 	if err != nil {
 		logger.Warnf("failed to insert key: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
-	ret, err := toWasmMemorySized(instanceContext, kp.Public().Encode())
+	ret, err := toWasmMemorySized(runtimeCtx, kp.Public().Encode())
 	if err != nil {
 		logger.Warnf("failed to allocate memory: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	logger.Debug("generated ed25519 keypair with public key: " + kp.Public().Hex())
-	return C.int32_t(ret)
+	return []wasmer.Value{wasmer.NewI32(int32(ret))}, nil
 }
 
 //export ext_crypto_ed25519_public_keys_version_1
-func ext_crypto_ed25519_public_keys_version_1(context unsafe.Pointer, keyTypeID C.int32_t) C.int64_t {
+func ext_crypto_ed25519_public_keys_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
+	runtimeCtx := env.(*runtime.Context)
+	memory := runtimeCtx.Memory.Data()
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	memory := instanceContext.Memory().Data()
-
+	keyTypeID := args[0].I32()
 	id := memory[keyTypeID : keyTypeID+4]
 
 	ks, err := runtimeCtx.Keystore.GetKeystore(id)
 	if err != nil {
 		logger.Warnf("error for id 0x%x: %s", id, err)
-		ret, _ := toWasmMemory(instanceContext, []byte{0})
-		return C.int64_t(ret)
+		ret, _ := toWasmMemory(runtimeCtx, []byte{0})
+		return []wasmer.Value{wasmer.NewI64(ret)}, nil
 	}
 
 	if ks.Type() != crypto.Ed25519Type && ks.Type() != crypto.UnknownType {
 		logger.Warnf(
 			"error for id 0x%x: keystore type is %s and not the expected ed25519",
 			id, ks.Type())
-		ret, _ := toWasmMemory(instanceContext, []byte{0})
-		return C.int64_t(ret)
+		ret, _ := toWasmMemory(runtimeCtx, []byte{0})
+		return []wasmer.Value{wasmer.NewI64(ret)}, nil
 	}
 
 	keys := ks.PublicKeys()
@@ -304,85 +218,90 @@ func ext_crypto_ed25519_public_keys_version_1(context unsafe.Pointer, keyTypeID 
 	prefix, err := scale.Marshal(big.NewInt(int64(len(keys))))
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
-		ret, _ := toWasmMemory(instanceContext, []byte{0})
-		return C.int64_t(ret)
+		ret, _ := toWasmMemory(runtimeCtx, []byte{0})
+		return []wasmer.Value{wasmer.NewI64(ret)}, nil
 	}
 
-	ret, err := toWasmMemory(instanceContext, append(prefix, encodedKeys...))
+	ret, err := toWasmMemory(runtimeCtx, append(prefix, encodedKeys...))
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
-		ret, _ = toWasmMemory(instanceContext, []byte{0})
-		return C.int64_t(ret)
+		ret, _ = toWasmMemory(runtimeCtx, []byte{0})
+		return []wasmer.Value{wasmer.NewI64(ret)}, nil
 	}
 
-	return C.int64_t(ret)
+	return []wasmer.Value{wasmer.NewI64(ret)}, nil
 }
 
 //export ext_crypto_ed25519_sign_version_1
-func ext_crypto_ed25519_sign_version_1(context unsafe.Pointer, keyTypeID, key C.int32_t, msg C.int64_t) C.int64_t {
+func ext_crypto_ed25519_sign_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	memory := instanceContext.Memory().Data()
+	runtimeCtx := env.(*runtime.Context)
+	memory := runtimeCtx.Memory.Data()
 
+	keyTypeID := args[0].I32()
+	key := args[1].I32()
+	msg := args[2].I64()
 	id := memory[keyTypeID : keyTypeID+4]
 
 	pubKeyData := memory[key : key+32]
 	pubKey, err := ed25519.NewPublicKey(pubKeyData)
 	if err != nil {
 		logger.Errorf("failed to get public keys: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
 	ks, err := runtimeCtx.Keystore.GetKeystore(id)
 	if err != nil {
 		logger.Warnf("error for id 0x%x: %s", id, err)
-		return mustToWasmMemoryOptionalNil(instanceContext)
+		return mustToWasmMemoryOptionalNil(runtimeCtx), nil
 	}
 
 	signingKey := ks.GetKeypair(pubKey)
 	if signingKey == nil {
 		logger.Error("could not find public key " + pubKey.Hex() + " in keystore")
-		ret, err := toWasmMemoryOptionalNil(instanceContext)
+		ret, err := toWasmMemoryOptionalNil(runtimeCtx)
 		if err != nil {
 			logger.Errorf("failed to allocate memory: %s", err)
-			return 0
+			return []wasmer.Value{wasmer.NewI64(0)}, nil
 		}
-		return ret
+		return ret, nil
 	}
 
-	sig, err := signingKey.Sign(asMemorySlice(instanceContext, msg))
+	sig, err := signingKey.Sign(asMemorySlice(runtimeCtx, msg))
 	if err != nil {
 		logger.Error("could not sign message")
 	}
 
-	ret, err := toWasmMemoryFixedSizeOptional(instanceContext, sig)
+	ret, err := toWasmMemoryFixedSizeOptional(runtimeCtx, sig)
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(ret)
+	return []wasmer.Value{wasmer.NewI64(ret)}, nil
 }
 
 //export ext_crypto_ed25519_verify_version_1
-func ext_crypto_ed25519_verify_version_1(context unsafe.Pointer, sig C.int32_t,
-	msg C.int64_t, key C.int32_t) C.int32_t {
+func ext_crypto_ed25519_verify_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	memory := instanceContext.Memory().Data()
-	sigVerifier := instanceContext.Data().(*runtime.Context).SigVerifier
+	runtimeCtx := env.(*runtime.Context)
+	memory := runtimeCtx.Memory.Data()
+	sigVerifier := runtimeCtx.SigVerifier
+
+	sig := args[0].I32()
+	msg := args[1].I64()
+	key := args[2].I32()
 
 	signature := memory[sig : sig+64]
-	message := asMemorySlice(instanceContext, msg)
+	message := asMemorySlice(runtimeCtx, msg)
 	pubKeyData := memory[key : key+32]
 
 	pubKey, err := ed25519.NewPublicKey(pubKeyData)
 	if err != nil {
 		logger.Error("failed to create public key")
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, err
 	}
 
 	if sigVerifier.IsStarted() {
@@ -393,23 +312,26 @@ func ext_crypto_ed25519_verify_version_1(context unsafe.Pointer, sig C.int32_t,
 			VerifyFunc: ed25519.VerifySignature,
 		}
 		sigVerifier.Add(&signature)
-		return 1
+		return []wasmer.Value{wasmer.NewI32(1)}, nil
 	}
 
 	if ok, err := pubKey.Verify(message, signature); err != nil || !ok {
 		logger.Error("failed to verify")
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, err
 	}
 
 	logger.Debug("verified ed25519 signature")
-	return 1
+	return []wasmer.Value{wasmer.NewI32(1)}, nil
 }
 
 //export ext_crypto_secp256k1_ecdsa_recover_version_1
-func ext_crypto_secp256k1_ecdsa_recover_version_1(context unsafe.Pointer, sig, msg C.int32_t) C.int64_t {
+func ext_crypto_secp256k1_ecdsa_recover_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	memory := instanceContext.Memory().Data()
+	instanceContext := env.(*runtime.Context)
+	memory := instanceContext.Memory.Data()
+
+	sig := args[0].I32()
+	msg := args[1].I32()
 
 	// msg must be the 32-byte hash of the message to be signed.
 	// sig must be a 65-byte compact ECDSA signature containing the
@@ -423,9 +345,9 @@ func ext_crypto_secp256k1_ecdsa_recover_version_1(context unsafe.Pointer, sig, m
 		ret, err := toWasmMemoryResultEmpty(instanceContext)
 		if err != nil {
 			logger.Errorf("failed to allocate memory: %s", err)
-			return 0
+			return []wasmer.Value{wasmer.NewI64(0)}, nil
 		}
-		return ret
+		return ret, nil
 	}
 
 	logger.Debugf(
@@ -435,25 +357,29 @@ func ext_crypto_secp256k1_ecdsa_recover_version_1(context unsafe.Pointer, sig, m
 	ret, err := toWasmMemoryResult(instanceContext, pub[1:])
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(ret)
+	return []wasmer.Value{wasmer.NewI64(ret)}, nil
 }
 
 //export ext_crypto_secp256k1_ecdsa_recover_version_2
-func ext_crypto_secp256k1_ecdsa_recover_version_2(context unsafe.Pointer, sig, msg C.int32_t) C.int64_t {
+func ext_crypto_secp256k1_ecdsa_recover_version_2(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	return ext_crypto_secp256k1_ecdsa_recover_version_1(context, sig, msg)
+	return ext_crypto_secp256k1_ecdsa_recover_version_1(env, args)
 }
 
 //export ext_crypto_ecdsa_verify_version_2
-func ext_crypto_ecdsa_verify_version_2(context unsafe.Pointer, sig C.int32_t, msg C.int64_t, key C.int32_t) C.int32_t {
+func ext_crypto_ecdsa_verify_version_2(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	memory := instanceContext.Memory().Data()
-	sigVerifier := instanceContext.Data().(*runtime.Context).SigVerifier
+	instanceContext := env.(*runtime.Context)
+	memory := instanceContext.Memory.Data()
+	sigVerifier := instanceContext.SigVerifier
+
+	sig := args[0].I32()
+	msg := args[1].I64()
+	key := args[2].I32()
 
 	message := asMemorySlice(instanceContext, msg)
 	signature := memory[sig : sig+64]
@@ -463,7 +389,7 @@ func ext_crypto_ecdsa_verify_version_2(context unsafe.Pointer, sig C.int32_t, ms
 	err := pub.Decode(pubKey)
 	if err != nil {
 		logger.Errorf("failed to decode public key: %s", err)
-		return C.int32_t(0)
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	logger.Debugf("pub=%s, message=0x%x, signature=0x%x",
@@ -472,7 +398,7 @@ func ext_crypto_ecdsa_verify_version_2(context unsafe.Pointer, sig C.int32_t, ms
 	hash, err := common.Blake2bHash(message)
 	if err != nil {
 		logger.Errorf("failed to hash message: %s", err)
-		return C.int32_t(0)
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	if sigVerifier.IsStarted() {
@@ -483,7 +409,7 @@ func ext_crypto_ecdsa_verify_version_2(context unsafe.Pointer, sig C.int32_t, ms
 			VerifyFunc: secp256k1.VerifySignature,
 		}
 		sigVerifier.Add(&signature)
-		return C.int32_t(1)
+		return []wasmer.Value{wasmer.NewI32(1)}, nil
 	}
 
 	ok, err := pub.Verify(hash[:], signature)
@@ -493,18 +419,22 @@ func ext_crypto_ecdsa_verify_version_2(context unsafe.Pointer, sig C.int32_t, ms
 			message += ": " + err.Error()
 		}
 		logger.Errorf(message)
-		return C.int32_t(0)
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	logger.Debug("validated signature")
-	return C.int32_t(1)
+	return []wasmer.Value{wasmer.NewI32(1)}, nil
 }
 
 //export ext_crypto_secp256k1_ecdsa_recover_compressed_version_1
-func ext_crypto_secp256k1_ecdsa_recover_compressed_version_1(context unsafe.Pointer, sig, msg C.int32_t) C.int64_t {
+func ext_crypto_secp256k1_ecdsa_recover_compressed_version_1(env interface{},
+	args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	memory := instanceContext.Memory().Data()
+	instanceContext := env.(*runtime.Context)
+	memory := instanceContext.Memory.Data()
+
+	sig := args[0].I32()
+	msg := args[1].I32()
 
 	// msg must be the 32-byte hash of the message to be signed.
 	// sig must be a 65-byte compact ECDSA signature containing the
@@ -515,7 +445,7 @@ func ext_crypto_secp256k1_ecdsa_recover_compressed_version_1(context unsafe.Poin
 	cpub, err := secp256k1.RecoverPublicKeyCompressed(message, signature)
 	if err != nil {
 		logger.Errorf("failed to recover public key: %s", err)
-		return mustToWasmMemoryResultEmpty(instanceContext)
+		return mustToWasmMemoryResultEmpty(instanceContext), nil
 	}
 
 	logger.Debugf(
@@ -525,25 +455,27 @@ func ext_crypto_secp256k1_ecdsa_recover_compressed_version_1(context unsafe.Poin
 	ret, err := toWasmMemoryResult(instanceContext, cpub)
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(ret)
+	return []wasmer.Value{wasmer.NewI64(ret)}, nil
 }
 
 //export ext_crypto_secp256k1_ecdsa_recover_compressed_version_2
-func ext_crypto_secp256k1_ecdsa_recover_compressed_version_2(context unsafe.Pointer, sig, msg C.int32_t) C.int64_t {
+func ext_crypto_secp256k1_ecdsa_recover_compressed_version_2(env interface{},
+	args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	return ext_crypto_secp256k1_ecdsa_recover_compressed_version_1(context, sig, msg)
+	return ext_crypto_secp256k1_ecdsa_recover_compressed_version_1(env, args)
 }
 
 //export ext_crypto_sr25519_generate_version_1
-func ext_crypto_sr25519_generate_version_1(context unsafe.Pointer, keyTypeID C.int32_t, seedSpan C.int64_t) C.int32_t {
+func ext_crypto_sr25519_generate_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	memory := instanceContext.Memory().Data()
+	instanceContext := env.(*runtime.Context)
+	memory := instanceContext.Memory.Data()
+	keyTypeID := args[0].I32()
+	seedSpan := args[1].I64()
 
 	id := memory[keyTypeID : keyTypeID+4]
 	seedBytes := asMemorySlice(instanceContext, seedSpan)
@@ -552,7 +484,7 @@ func ext_crypto_sr25519_generate_version_1(context unsafe.Pointer, keyTypeID C.i
 	err := scale.Unmarshal(seedBytes, &seed)
 	if err != nil {
 		logger.Warnf("cannot generate key: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	var kp KeyPair
@@ -567,43 +499,49 @@ func ext_crypto_sr25519_generate_version_1(context unsafe.Pointer, keyTypeID C.i
 		panic(err)
 	}
 
-	ks, err := runtimeCtx.Keystore.GetKeystore(id)
+	ks, err := instanceContext.Keystore.GetKeystore(id)
 	if err != nil {
 		logger.Warnf("error for id "+common.BytesToHex(id)+": %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	err = ks.Insert(kp)
 	if err != nil {
 		logger.Warnf("failed to insert key: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	ret, err := toWasmMemorySized(instanceContext, kp.Public().Encode())
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
+	// TODO dont think this is safe
+	val := int32(ret)
+
 	logger.Debug("generated sr25519 keypair with public key: " + kp.Public().Hex())
-	return C.int32_t(ret)
+	return []wasmer.Value{wasmer.NewI32(val)}, nil
 }
 
 //export ext_crypto_sr25519_public_keys_version_1
-func ext_crypto_sr25519_public_keys_version_1(context unsafe.Pointer, keyTypeID C.int32_t) C.int64_t {
+func ext_crypto_sr25519_public_keys_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	memory := instanceContext.Memory().Data()
+	instanceContext := env.(*runtime.Context)
+	memory := instanceContext.Memory.Data()
+	keyTypeID, ok := args[0].Unwrap().(int32)
+	if !ok {
+		panic("keyTypeID is not int32")
+	}
 
 	id := memory[keyTypeID : keyTypeID+4]
 
-	ks, err := runtimeCtx.Keystore.GetKeystore(id)
+	ks, err := instanceContext.Keystore.GetKeystore(id)
 	if err != nil {
 		logger.Warnf("error for id "+common.BytesToHex(id)+": %s", err)
 		ret, _ := toWasmMemory(instanceContext, []byte{0})
-		return C.int64_t(ret)
+		return []wasmer.Value{wasmer.NewI64(ret)}, nil
 	}
 
 	if ks.Type() != crypto.Sr25519Type && ks.Type() != crypto.UnknownType {
@@ -611,7 +549,7 @@ func ext_crypto_sr25519_public_keys_version_1(context unsafe.Pointer, keyTypeID 
 			"keystore type for id 0x%x is %s and not expected sr25519",
 			id, ks.Type())
 		ret, _ := toWasmMemory(instanceContext, []byte{0})
-		return C.int64_t(ret)
+		return []wasmer.Value{wasmer.NewI64(ret)}, nil
 	}
 
 	keys := ks.PublicKeys()
@@ -625,71 +563,78 @@ func ext_crypto_sr25519_public_keys_version_1(context unsafe.Pointer, keyTypeID 
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
 		ret, _ := toWasmMemory(instanceContext, []byte{0})
-		return C.int64_t(ret)
+		return []wasmer.Value{wasmer.NewI64(ret)}, nil
 	}
 
 	ret, err := toWasmMemory(instanceContext, append(prefix, encodedKeys...))
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
 		ret, _ = toWasmMemory(instanceContext, []byte{0})
-		return C.int64_t(ret)
+		return []wasmer.Value{wasmer.NewI64(ret)}, nil
 	}
 
-	return C.int64_t(ret)
+	return []wasmer.Value{wasmer.NewI64(ret)}, nil
 }
 
 //export ext_crypto_sr25519_sign_version_1
-func ext_crypto_sr25519_sign_version_1(context unsafe.Pointer, keyTypeID, key C.int32_t, msg C.int64_t) C.int64_t {
+func ext_crypto_sr25519_sign_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	memory := instanceContext.Memory().Data()
+	runtimeCtx := env.(*runtime.Context)
+	memory := runtimeCtx.Memory.Data()
+	keyTypeID := args[0].I32()
+	key := args[1].I32()
+	msg := args[2].I64()
+
+	emptyRet, _ := toWasmMemoryOptional(runtimeCtx, nil)
 
 	id := memory[keyTypeID : keyTypeID+4]
 
 	ks, err := runtimeCtx.Keystore.GetKeystore(id)
 	if err != nil {
 		logger.Warnf("error for id 0x%x: %s", id, err)
-		return mustToWasmMemoryOptionalNil(instanceContext)
+		return []wasmer.Value{wasmer.NewI64(emptyRet)}, nil
 	}
 
 	var ret int64
 	pubKey, err := sr25519.NewPublicKey(memory[key : key+32])
 	if err != nil {
 		logger.Errorf("failed to get public key: %s", err)
-		return mustToWasmMemoryOptionalNil(instanceContext)
+		return []wasmer.Value{wasmer.NewI64(emptyRet)}, nil
 	}
 
 	signingKey := ks.GetKeypair(pubKey)
 	if signingKey == nil {
 		logger.Error("could not find public key " + pubKey.Hex() + " in keystore")
-		return mustToWasmMemoryOptionalNil(instanceContext)
+		return []wasmer.Value{wasmer.NewI64(emptyRet)}, nil
 	}
 
-	msgData := asMemorySlice(instanceContext, msg)
+	msgData := asMemorySlice(runtimeCtx, msg)
 	sig, err := signingKey.Sign(msgData)
 	if err != nil {
 		logger.Errorf("could not sign message: %s", err)
-		return mustToWasmMemoryOptionalNil(instanceContext)
+		return []wasmer.Value{wasmer.NewI64(emptyRet)}, nil
 	}
 
-	ret, err = toWasmMemoryFixedSizeOptional(instanceContext, sig)
+	ret, err = toWasmMemoryFixedSizeOptional(runtimeCtx, sig)
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
-		return mustToWasmMemoryOptionalNil(instanceContext)
+		return []wasmer.Value{wasmer.NewI64(emptyRet)}, nil
 	}
 
-	return C.int64_t(ret)
+	return []wasmer.Value{wasmer.NewI64(ret)}, nil
 }
 
 //export ext_crypto_sr25519_verify_version_1
-func ext_crypto_sr25519_verify_version_1(context unsafe.Pointer, sig C.int32_t,
-	msg C.int64_t, key C.int32_t) C.int32_t {
+func ext_crypto_sr25519_verify_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	memory := instanceContext.Memory().Data()
-	sigVerifier := instanceContext.Data().(*runtime.Context).SigVerifier
+	instanceContext := env.(*runtime.Context)
+	memory := instanceContext.Memory.Data()
+	sigVerifier := instanceContext.SigVerifier
+
+	sig := args[0].I32()
+	msg := args[1].I64()
+	key := args[2].I32()
 
 	message := asMemorySlice(instanceContext, msg)
 	signature := memory[sig : sig+64]
@@ -697,7 +642,7 @@ func ext_crypto_sr25519_verify_version_1(context unsafe.Pointer, sig C.int32_t,
 	pub, err := sr25519.NewPublicKey(memory[key : key+32])
 	if err != nil {
 		logger.Error("invalid sr25519 public key")
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, err
 	}
 
 	logger.Debugf(
@@ -712,7 +657,7 @@ func ext_crypto_sr25519_verify_version_1(context unsafe.Pointer, sig C.int32_t,
 			VerifyFunc: sr25519.VerifySignature,
 		}
 		sigVerifier.Add(&signature)
-		return 1
+		return []wasmer.Value{wasmer.NewI32(1)}, nil
 	}
 
 	ok, err := pub.VerifyDeprecated(message, signature)
@@ -723,21 +668,24 @@ func ext_crypto_sr25519_verify_version_1(context unsafe.Pointer, sig C.int32_t,
 		}
 		logger.Debugf(message)
 		// this fails at block 3876, which seems to be expected, based on discussions
-		return 1
+		return []wasmer.Value{wasmer.NewI32(1)}, nil
 	}
 
 	logger.Debug("verified sr25519 signature")
-	return 1
+	return []wasmer.Value{wasmer.NewI32(1)}, nil
 }
 
 //export ext_crypto_sr25519_verify_version_2
-func ext_crypto_sr25519_verify_version_2(context unsafe.Pointer, sig C.int32_t,
-	msg C.int64_t, key C.int32_t) C.int32_t {
+func ext_crypto_sr25519_verify_version_2(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	memory := instanceContext.Memory().Data()
-	sigVerifier := instanceContext.Data().(*runtime.Context).SigVerifier
+	instanceContext := env.(*runtime.Context)
+	memory := instanceContext.Memory.Data()
+	sigVerifier := instanceContext.SigVerifier
+
+	sig := args[0].I32()
+	msg := args[1].I64()
+	key := args[2].I32()
 
 	message := asMemorySlice(instanceContext, msg)
 	signature := memory[sig : sig+64]
@@ -745,7 +693,7 @@ func ext_crypto_sr25519_verify_version_2(context unsafe.Pointer, sig C.int32_t,
 	pub, err := sr25519.NewPublicKey(memory[key : key+32])
 	if err != nil {
 		logger.Error("invalid sr25519 public key")
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, err
 	}
 
 	logger.Debugf(
@@ -760,7 +708,7 @@ func ext_crypto_sr25519_verify_version_2(context unsafe.Pointer, sig C.int32_t,
 			VerifyFunc: sr25519.VerifySignature,
 		}
 		sigVerifier.Add(&signature)
-		return 1
+		return []wasmer.Value{wasmer.NewI32(1)}, nil
 	}
 
 	ok, err := pub.Verify(message, signature)
@@ -770,38 +718,40 @@ func ext_crypto_sr25519_verify_version_2(context unsafe.Pointer, sig C.int32_t,
 			message += ": " + err.Error()
 		}
 		logger.Errorf(message)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	logger.Debug("validated signature")
-	return C.int32_t(1)
+	return []wasmer.Value{wasmer.NewI32(1)}, nil
 }
 
 //export ext_crypto_start_batch_verify_version_1
-func ext_crypto_start_batch_verify_version_1(context unsafe.Pointer) {
+func ext_crypto_start_batch_verify_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
 	// TODO: fix and re-enable signature verification (#1405)
 	// beginBatchVerify(context)
+	return nil, nil
 }
 
 //export ext_crypto_finish_batch_verify_version_1
-func ext_crypto_finish_batch_verify_version_1(context unsafe.Pointer) C.int32_t {
+func ext_crypto_finish_batch_verify_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
 	// TODO: fix and re-enable signature verification (#1405)
 	// return finishBatchVerify(context)
-	return 1
+	return []wasmer.Value{wasmer.NewI32(1)}, nil
 }
 
 //export ext_trie_blake2_256_root_version_1
-func ext_trie_blake2_256_root_version_1(context unsafe.Pointer, dataSpan C.int64_t) C.int32_t {
+func ext_trie_blake2_256_root_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	memory := instanceContext.Memory().Data()
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	data := asMemorySlice(instanceContext, dataSpan)
+	runtimeCtx := env.(*runtime.Context)
+	memory := runtimeCtx.Memory.Data()
+	dataSpan := args[0].I64()
+
+	data := asMemorySlice(runtimeCtx, dataSpan)
 
 	t := trie.NewEmptyTrie()
 
@@ -813,7 +763,7 @@ func ext_trie_blake2_256_root_version_1(context unsafe.Pointer, dataSpan C.int64
 	var kvs []kv
 	if err := scale.Unmarshal(data, &kvs); err != nil {
 		logger.Errorf("failed scale decoding data: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	for _, kv := range kvs {
@@ -821,7 +771,7 @@ func ext_trie_blake2_256_root_version_1(context unsafe.Pointer, dataSpan C.int64
 		if err != nil {
 			logger.Errorf("failed putting key 0x%x and value 0x%x into trie: %s",
 				kv.Key, kv.Value, err)
-			return 0
+			return []wasmer.Value{wasmer.NewI32(0)}, nil
 		}
 	}
 
@@ -829,42 +779,44 @@ func ext_trie_blake2_256_root_version_1(context unsafe.Pointer, dataSpan C.int64
 	ptr, err := runtimeCtx.Allocator.Allocate(32)
 	if err != nil {
 		logger.Errorf("failed allocating: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	hash, err := t.Hash()
 	if err != nil {
 		logger.Errorf("failed computing trie Merkle root hash: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	logger.Debugf("root hash is %s", hash)
 	copy(memory[ptr:ptr+32], hash[:])
-	return C.int32_t(ptr)
+	// TODO is this safe?
+	return []wasmer.Value{wasmer.NewI32(int32(ptr))}, nil
 }
 
 //export ext_trie_blake2_256_ordered_root_version_1
-func ext_trie_blake2_256_ordered_root_version_1(context unsafe.Pointer, dataSpan C.int64_t) C.int32_t {
+func ext_trie_blake2_256_ordered_root_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	memory := instanceContext.Memory().Data()
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	data := asMemorySlice(instanceContext, dataSpan)
+	runtimeCtx := env.(*runtime.Context)
+	memory := runtimeCtx.Memory.Data()
+	dataSpan := args[0].I64()
+
+	data := asMemorySlice(runtimeCtx, dataSpan)
 
 	t := trie.NewEmptyTrie()
 	var values [][]byte
 	err := scale.Unmarshal(data, &values)
 	if err != nil {
 		logger.Errorf("failed scale decoding data: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	for i, value := range values {
 		key, err := scale.Marshal(big.NewInt(int64(i)))
 		if err != nil {
 			logger.Errorf("failed scale encoding value index %d: %s", i, err)
-			return 0
+			return []wasmer.Value{wasmer.NewI32(0)}, nil
 		}
 		logger.Tracef(
 			"put key=0x%x and value=0x%x",
@@ -874,7 +826,7 @@ func ext_trie_blake2_256_ordered_root_version_1(context unsafe.Pointer, dataSpan
 		if err != nil {
 			logger.Errorf("failed putting key 0x%x and value 0x%x into trie: %s",
 				key, value, err)
-			return 0
+			return []wasmer.Value{wasmer.NewI32(0)}, nil
 		}
 	}
 
@@ -882,93 +834,103 @@ func ext_trie_blake2_256_ordered_root_version_1(context unsafe.Pointer, dataSpan
 	ptr, err := runtimeCtx.Allocator.Allocate(32)
 	if err != nil {
 		logger.Errorf("failed allocating: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	hash, err := t.Hash()
 	if err != nil {
 		logger.Errorf("failed computing trie Merkle root hash: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	logger.Debugf("root hash is %s", hash)
 	copy(memory[ptr:ptr+32], hash[:])
-	return C.int32_t(ptr)
+	// TODO is this safe cast
+	return []wasmer.Value{wasmer.NewI32(int32(ptr))}, nil
 }
 
 //export ext_trie_blake2_256_ordered_root_version_2
-func ext_trie_blake2_256_ordered_root_version_2(context unsafe.Pointer,
-	dataSpan C.int64_t, version C.int32_t) C.int32_t {
+func ext_trie_blake2_256_ordered_root_version_2(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	// TODO: update to use state trie version 1 (#2418)
-	return ext_trie_blake2_256_ordered_root_version_1(context, dataSpan)
+	return ext_trie_blake2_256_ordered_root_version_1(env, args)
 }
 
 //export ext_trie_blake2_256_verify_proof_version_1
-func ext_trie_blake2_256_verify_proof_version_1(context unsafe.Pointer,
-	rootSpan C.int32_t, proofSpan, keySpan, valueSpan C.int64_t) C.int32_t {
+func ext_trie_blake2_256_verify_proof_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
+	instanceContext := env.(*runtime.Context)
+	rootSpan := args[0].I32()
+	proofSpan := args[1].I64()
+	keySpan := args[2].I64()
+	valueSpan := args[3].I64()
 
 	toDecProofs := asMemorySlice(instanceContext, proofSpan)
 	var encodedProofNodes [][]byte
 	err := scale.Unmarshal(toDecProofs, &encodedProofNodes)
 	if err != nil {
 		logger.Errorf("failed scale decoding proof data: %s", err)
-		return C.int32_t(0)
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	key := asMemorySlice(instanceContext, keySpan)
 	value := asMemorySlice(instanceContext, valueSpan)
 
-	mem := instanceContext.Memory().Data()
+	mem := instanceContext.Memory.Data()
 	trieRoot := mem[rootSpan : rootSpan+32]
 
 	err = proof.Verify(encodedProofNodes, trieRoot, key, value)
 	if err != nil {
 		logger.Errorf("failed proof verification: %s", err)
-		return C.int32_t(0)
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
-	return C.int32_t(1)
+	return []wasmer.Value{wasmer.NewI32(1)}, nil
 }
 
 //export ext_misc_print_hex_version_1
-func ext_misc_print_hex_version_1(context unsafe.Pointer, dataSpan C.int64_t) {
+func ext_misc_print_hex_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	data := asMemorySlice(instanceContext, dataSpan)
+	ctx := env.(*runtime.Context)
+	dataSpan := args[0].I64()
+	data := asMemorySlice(ctx, dataSpan)
 	logger.Debugf("data: 0x%x", data)
+	return nil, nil
 }
 
 //export ext_misc_print_num_version_1
-func ext_misc_print_num_version_1(_ unsafe.Pointer, data C.int64_t) {
+func ext_misc_print_num_version_1(_ interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-
-	logger.Debugf("num: %d", int64(data))
+	data := args[0].I64()
+	logger.Debugf("num: %d", data)
+	return nil, nil
 }
 
 //export ext_misc_print_utf8_version_1
-func ext_misc_print_utf8_version_1(context unsafe.Pointer, dataSpan C.int64_t) {
+func ext_misc_print_utf8_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	data := asMemorySlice(instanceContext, dataSpan)
+	ctx := env.(*runtime.Context)
+	dataSpan := args[0].I64()
+	data := asMemorySlice(ctx, dataSpan)
 	logger.Debug("utf8: " + string(data))
+	return nil, nil
 }
 
 //export ext_misc_runtime_version_version_1
-func ext_misc_runtime_version_version_1(context unsafe.Pointer, dataSpan C.int64_t) C.int64_t {
+func ext_misc_runtime_version_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
+	instanceContext := env.(*runtime.Context)
+	dataSpan := args[0].I64()
+	//data := asMemorySlice(ctx, dataSpan)
 	code := asMemorySlice(instanceContext, dataSpan)
 
 	version, err := GetRuntimeVersion(code)
 	if err != nil {
 		logger.Errorf("failed to get runtime version: %s", err)
-		return mustToWasmMemoryOptionalNil(instanceContext)
+		return mustToWasmMemoryOptionalNil(instanceContext), nil
 	}
 
 	// Note the encoding contains all the latest Core_version fields as defined in
@@ -981,36 +943,40 @@ func ext_misc_runtime_version_version_1(context unsafe.Pointer, dataSpan C.int64
 	encodedData, err := scale.Marshal(version)
 	if err != nil {
 		logger.Errorf("failed to encode result: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
 	out, err := toWasmMemoryOptional(instanceContext, encodedData)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(out)
+	return []wasmer.Value{wasmer.NewI64(out)}, nil
 }
 
 //export ext_default_child_storage_read_version_1
-func ext_default_child_storage_read_version_1(context unsafe.Pointer,
-	childStorageKey, key, valueOut C.int64_t, offset C.int32_t) C.int64_t {
+func ext_default_child_storage_read_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	storage := instanceContext.Data().(*runtime.Context).Storage
-	memory := instanceContext.Memory().Data()
+	instanceContext := env.(*runtime.Context)
+	memory := instanceContext.Memory.Data()
+	storage := instanceContext.Storage
+
+	childStorageKey := args[0].I64()
+	key := args[1].I64()
+	valueOut := args[2].I64()
+	offset := args[3].I32()
 
 	keyToChild := asMemorySlice(instanceContext, childStorageKey)
 	keyBytes := asMemorySlice(instanceContext, key)
 	value, err := storage.GetChildStorage(keyToChild, keyBytes)
 	if err != nil {
 		logger.Errorf("failed to get child storage: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	valueBuf, valueLen := splitPointerSize(int64(valueOut))
+	valueBuf, valueLen := splitPointerSize(valueOut)
 	copy(memory[valueBuf:valueBuf+valueLen], value[offset:])
 
 	size := uint32(len(value[offset:]))
@@ -1020,19 +986,20 @@ func ext_default_child_storage_read_version_1(context unsafe.Pointer,
 	sizeSpan, err := toWasmMemoryOptional(instanceContext, sizeBuf)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(sizeSpan)
+	return []wasmer.Value{wasmer.NewI64(sizeSpan)}, nil
 }
 
 //export ext_default_child_storage_clear_version_1
-func ext_default_child_storage_clear_version_1(context unsafe.Pointer, childStorageKey, keySpan C.int64_t) {
+func ext_default_child_storage_clear_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
+	childStorageKey := args[0].I64()
+	keySpan := args[1].I64()
 
 	keyToChild := asMemorySlice(instanceContext, childStorageKey)
 	key := asMemorySlice(instanceContext, keySpan)
@@ -1041,15 +1008,17 @@ func ext_default_child_storage_clear_version_1(context unsafe.Pointer, childStor
 	if err != nil {
 		logger.Errorf("failed to clear child storage: %s", err)
 	}
+	return nil, nil
 }
 
 //export ext_default_child_storage_clear_prefix_version_1
-func ext_default_child_storage_clear_prefix_version_1(context unsafe.Pointer, childStorageKey, prefixSpan C.int64_t) {
+func ext_default_child_storage_clear_prefix_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
+	childStorageKey := args[0].I64()
+	prefixSpan := args[1].I64()
 
 	keyToChild := asMemorySlice(instanceContext, childStorageKey)
 	prefix := asMemorySlice(instanceContext, prefixSpan)
@@ -1058,114 +1027,132 @@ func ext_default_child_storage_clear_prefix_version_1(context unsafe.Pointer, ch
 	if err != nil {
 		logger.Errorf("failed to clear prefix in child: %s", err)
 	}
+	return nil, nil
 }
 
 //export ext_default_child_storage_exists_version_1
-func ext_default_child_storage_exists_version_1(context unsafe.Pointer,
-	childStorageKey, key C.int64_t) C.int32_t {
+func ext_default_child_storage_exists_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	storage := instanceContext.Data().(*runtime.Context).Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
+
+	childStorageKey := args[0].I64()
+	key := args[1].I64()
 
 	keyToChild := asMemorySlice(instanceContext, childStorageKey)
 	keyBytes := asMemorySlice(instanceContext, key)
 	child, err := storage.GetChildStorage(keyToChild, keyBytes)
 	if err != nil {
 		logger.Errorf("failed to get child from child storage: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 	if child != nil {
-		return 1
+		return []wasmer.Value{wasmer.NewI32(1)}, nil
 	}
-	return 0
+	return []wasmer.Value{wasmer.NewI32(0)}, nil
 }
 
 //export ext_default_child_storage_get_version_1
-func ext_default_child_storage_get_version_1(context unsafe.Pointer, childStorageKey, key C.int64_t) C.int64_t {
+func ext_default_child_storage_get_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	storage := instanceContext.Data().(*runtime.Context).Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
+
+	childStorageKey := args[0].I64()
+	key := args[1].I64()
 
 	keyToChild := asMemorySlice(instanceContext, childStorageKey)
 	keyBytes := asMemorySlice(instanceContext, key)
 	child, err := storage.GetChildStorage(keyToChild, keyBytes)
 	if err != nil {
 		logger.Errorf("failed to get child from child storage: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
 	value, err := toWasmMemoryOptional(instanceContext, child)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(value)
+	return []wasmer.Value{wasmer.NewI64(value)}, nil
 }
 
 //export ext_default_child_storage_next_key_version_1
-func ext_default_child_storage_next_key_version_1(context unsafe.Pointer, childStorageKey, key C.int64_t) C.int64_t {
+func ext_default_child_storage_next_key_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	storage := instanceContext.Data().(*runtime.Context).Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
+
+	childStorageKey, ok := args[0].Unwrap().(int64)
+	if !ok {
+		panic("childStorageKey is not int64")
+	}
+	key, ok := args[1].Unwrap().(int64)
+	if !ok {
+		panic("key is not int64")
+	}
 
 	keyToChild := asMemorySlice(instanceContext, childStorageKey)
 	keyBytes := asMemorySlice(instanceContext, key)
 	child, err := storage.GetChildNextKey(keyToChild, keyBytes)
 	if err != nil {
 		logger.Errorf("failed to get child's next key: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
 	value, err := toWasmMemoryOptional(instanceContext, child)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(value)
+	return []wasmer.Value{wasmer.NewI64(value)}, nil
 }
 
 //export ext_default_child_storage_root_version_1
-func ext_default_child_storage_root_version_1(context unsafe.Pointer,
-	childStorageKey C.int64_t) (ptrSize C.int64_t) {
+func ext_default_child_storage_root_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	storage := instanceContext.Data().(*runtime.Context).Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
+
+	childStorageKey := args[0].I64()
 
 	child, err := storage.GetChild(asMemorySlice(instanceContext, childStorageKey))
 	if err != nil {
 		logger.Errorf("failed to retrieve child: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
 	childRoot, err := child.Hash()
 	if err != nil {
 		logger.Errorf("failed to encode child root: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
 	root, err := toWasmMemoryOptional(instanceContext, childRoot[:])
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(root)
+	return []wasmer.Value{wasmer.NewI64(root)}, nil
 }
 
 //export ext_default_child_storage_set_version_1
-func ext_default_child_storage_set_version_1(context unsafe.Pointer,
-	childStorageKeySpan, keySpan, valueSpan C.int64_t) {
+func ext_default_child_storage_set_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
+
+	childStorageKeySpan := args[0].I64()
+	keySpan := args[1].I64()
+	valueSpan := args[2].I64()
 
 	childStorageKey := asMemorySlice(instanceContext, childStorageKeySpan)
 	key := asMemorySlice(instanceContext, keySpan)
@@ -1177,31 +1164,34 @@ func ext_default_child_storage_set_version_1(context unsafe.Pointer,
 	err := storage.SetChildStorage(childStorageKey, key, cp)
 	if err != nil {
 		logger.Errorf("failed to set value in child storage: %s", err)
-		return
+		return nil, nil
 	}
+	return nil, nil
 }
 
 //export ext_default_child_storage_storage_kill_version_1
-func ext_default_child_storage_storage_kill_version_1(context unsafe.Pointer, childStorageKeySpan C.int64_t) {
+func ext_default_child_storage_storage_kill_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
+	childStorageKeySpan := args[0].I64()
 
 	childStorageKey := asMemorySlice(instanceContext, childStorageKeySpan)
 	err := storage.DeleteChild(childStorageKey)
 	panicOnError(err)
+	return nil, nil
 }
 
 //export ext_default_child_storage_storage_kill_version_2
-func ext_default_child_storage_storage_kill_version_2(context unsafe.Pointer,
-	childStorageKeySpan, lim C.int64_t) (allDeleted C.int32_t) {
+func ext_default_child_storage_storage_kill_version_2(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
+	childStorageKeySpan := args[0].I64()
+	lim := args[1].I64()
+
 	childStorageKey := asMemorySlice(instanceContext, childStorageKeySpan)
 
 	limitBytes := asMemorySlice(instanceContext, lim)
@@ -1210,7 +1200,7 @@ func ext_default_child_storage_storage_kill_version_2(context unsafe.Pointer,
 	err := scale.Unmarshal(limitBytes, &limit)
 	if err != nil {
 		logger.Warnf("cannot generate limit: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(0)}, nil
 	}
 
 	_, all, err := storage.DeleteChildLimit(childStorageKey, limit)
@@ -1219,10 +1209,10 @@ func ext_default_child_storage_storage_kill_version_2(context unsafe.Pointer,
 	}
 
 	if all {
-		return 1
+		return []wasmer.Value{wasmer.NewI32(1)}, nil
 	}
 
-	return 0
+	return []wasmer.Value{wasmer.NewI32(0)}, nil
 }
 
 type noneRemain uint32
@@ -1236,12 +1226,12 @@ func (someRemain) Index() uint       { return 1 }
 func (sr someRemain) String() string { return fmt.Sprintf("someRemain(%d)", sr) }
 
 //export ext_default_child_storage_storage_kill_version_3
-func ext_default_child_storage_storage_kill_version_3(context unsafe.Pointer,
-	childStorageKeySpan, lim C.int64_t) (pointerSize C.int64_t) {
+func ext_default_child_storage_storage_kill_version_3(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
+	childStorageKeySpan := args[0].I64()
+	lim := args[1].I64()
 	childStorageKey := asMemorySlice(instanceContext, childStorageKeySpan)
 
 	limitBytes := asMemorySlice(instanceContext, lim)
@@ -1255,7 +1245,7 @@ func ext_default_child_storage_storage_kill_version_3(context unsafe.Pointer,
 	deleted, all, err := storage.DeleteChildLimit(childStorageKey, limit)
 	if err != nil {
 		logger.Warnf("cannot get child storage: %s", err)
-		return C.int64_t(0)
+		return []wasmer.Value{wasmer.NewI64(0)}, err
 	}
 
 	vdt, err := scale.NewVaryingDataType(noneRemain(0), someRemain(0))
@@ -1270,43 +1260,50 @@ func ext_default_child_storage_storage_kill_version_3(context unsafe.Pointer,
 	}
 	if err != nil {
 		logger.Warnf("cannot set varying data type: %s", err)
-		return C.int64_t(0)
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
 	encoded, err := scale.Marshal(vdt)
 	if err != nil {
 		logger.Warnf("problem marshalling varying data type: %s", err)
-		return C.int64_t(0)
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
 	out, err := toWasmMemoryOptional(instanceContext, encoded)
 	if err != nil {
 		logger.Warnf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(out)
+	return []wasmer.Value{wasmer.NewI64(out)}, nil
 }
 
 //export ext_allocator_free_version_1
-func ext_allocator_free_version_1(context unsafe.Pointer, addr C.int32_t) {
+func ext_allocator_free_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
+	runtimeCtx := env.(*runtime.Context)
+	addr, ok := args[0].Unwrap().(int32)
+	if !ok {
+		logger.Criticalf("[ext_allocator_free_version_1]", "error", "addr cannot be converted to int32")
+	}
 
 	// Deallocate memory
 	err := runtimeCtx.Allocator.Deallocate(uint32(addr))
 	if err != nil {
 		logger.Errorf("failed to free memory: %s", err)
 	}
+	return nil, nil
 }
 
 //export ext_allocator_malloc_version_1
-func ext_allocator_malloc_version_1(context unsafe.Pointer, size C.int32_t) C.int32_t {
+func ext_allocator_malloc_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
+	size, ok := args[0].Unwrap().(int32)
+	if !ok {
+		logger.Criticalf("[ext_allocator_malloc_version_1]", "error", "addr cannot be converted to int32")
+	}
 	logger.Tracef("executing with size %d...", int64(size))
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
+	ctx := env.(*runtime.Context)
 
 	// Allocate memory
 	res, err := ctx.Allocator.Allocate(uint32(size))
@@ -1315,20 +1312,20 @@ func ext_allocator_malloc_version_1(context unsafe.Pointer, size C.int32_t) C.in
 		panic(err)
 	}
 
-	return C.int32_t(res)
+	return []wasmer.Value{wasmer.NewI32(int32(res))}, nil
 }
 
 //export ext_hashing_blake2_128_version_1
-func ext_hashing_blake2_128_version_1(context unsafe.Pointer, dataSpan C.int64_t) C.int32_t {
+func ext_hashing_blake2_128_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-
+	instanceContext := env.(*runtime.Context)
+	dataSpan := args[0].I64()
 	data := asMemorySlice(instanceContext, dataSpan)
 
 	hash, err := common.Blake2b128(data)
 	if err != nil {
 		logger.Errorf("failed hashing data: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
 	logger.Debugf(
@@ -1338,23 +1335,23 @@ func ext_hashing_blake2_128_version_1(context unsafe.Pointer, dataSpan C.int64_t
 	out, err := toWasmMemorySized(instanceContext, hash)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
-	return C.int32_t(out)
+	return []wasmer.Value{wasmer.NewI32(int32(out))}, nil
 }
 
 //export ext_hashing_blake2_256_version_1
-func ext_hashing_blake2_256_version_1(context unsafe.Pointer, dataSpan C.int64_t) C.int32_t {
+func ext_hashing_blake2_256_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-
+	instanceContext := env.(*runtime.Context)
+	dataSpan := args[0].I64()
 	data := asMemorySlice(instanceContext, dataSpan)
 
 	hash, err := common.Blake2bHash(data)
 	if err != nil {
 		logger.Errorf("failed hashing data: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
 	logger.Debugf("data 0x%x has hash %s", data, hash)
@@ -1362,23 +1359,23 @@ func ext_hashing_blake2_256_version_1(context unsafe.Pointer, dataSpan C.int64_t
 	out, err := toWasmMemorySized(instanceContext, hash[:])
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
-	return C.int32_t(out)
+	return []wasmer.Value{wasmer.NewI32(int32(out))}, nil
 }
 
 //export ext_hashing_keccak_256_version_1
-func ext_hashing_keccak_256_version_1(context unsafe.Pointer, dataSpan C.int64_t) C.int32_t {
+func ext_hashing_keccak_256_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-
+	instanceContext := env.(*runtime.Context)
+	dataSpan := args[0].I64()
 	data := asMemorySlice(instanceContext, dataSpan)
 
 	hash, err := common.Keccak256(data)
 	if err != nil {
 		logger.Errorf("failed hashing data: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
 	logger.Debugf("data 0x%x has hash %s", data, hash)
@@ -1386,17 +1383,17 @@ func ext_hashing_keccak_256_version_1(context unsafe.Pointer, dataSpan C.int64_t
 	out, err := toWasmMemorySized(instanceContext, hash[:])
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
-	return C.int32_t(out)
+	return []wasmer.Value{wasmer.NewI32(int32(out))}, nil
 }
 
 //export ext_hashing_sha2_256_version_1
-func ext_hashing_sha2_256_version_1(context unsafe.Pointer, dataSpan C.int64_t) C.int32_t {
+func ext_hashing_sha2_256_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-
+	instanceContext := env.(*runtime.Context)
+	dataSpan := args[0].I64()
 	data := asMemorySlice(instanceContext, dataSpan)
 	hash := common.Sha256(data)
 
@@ -1405,23 +1402,23 @@ func ext_hashing_sha2_256_version_1(context unsafe.Pointer, dataSpan C.int64_t) 
 	out, err := toWasmMemorySized(instanceContext, hash[:])
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
-	return C.int32_t(out)
+	return []wasmer.Value{wasmer.NewI32(int32(out))}, nil
 }
 
 //export ext_hashing_twox_256_version_1
-func ext_hashing_twox_256_version_1(context unsafe.Pointer, dataSpan C.int64_t) C.int32_t {
+func ext_hashing_twox_256_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-
+	instanceContext := env.(*runtime.Context)
+	dataSpan := args[0].I64()
 	data := asMemorySlice(instanceContext, dataSpan)
 
 	hash, err := common.Twox256(data)
 	if err != nil {
 		logger.Errorf("failed hashing data: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
 	logger.Debugf("data 0x%x has hash %s", data, hash)
@@ -1429,22 +1426,23 @@ func ext_hashing_twox_256_version_1(context unsafe.Pointer, dataSpan C.int64_t) 
 	out, err := toWasmMemorySized(instanceContext, hash[:])
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
-	return C.int32_t(out)
+	return []wasmer.Value{wasmer.NewI32(int32(out))}, nil
 }
 
 //export ext_hashing_twox_128_version_1
-func ext_hashing_twox_128_version_1(context unsafe.Pointer, dataSpan C.int64_t) C.int32_t {
+func ext_hashing_twox_128_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
+	instanceContext := env.(*runtime.Context)
+	dataSpan := args[0].I64()
 	data := asMemorySlice(instanceContext, dataSpan)
 
 	hash, err := common.Twox128Hash(data)
 	if err != nil {
 		logger.Errorf("failed hashing data: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
 	logger.Debugf(
@@ -1454,23 +1452,23 @@ func ext_hashing_twox_128_version_1(context unsafe.Pointer, dataSpan C.int64_t) 
 	out, err := toWasmMemorySized(instanceContext, hash)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
-	return C.int32_t(out)
+	return []wasmer.Value{wasmer.NewI32(int32(out))}, nil
 }
 
 //export ext_hashing_twox_64_version_1
-func ext_hashing_twox_64_version_1(context unsafe.Pointer, dataSpan C.int64_t) C.int32_t {
+func ext_hashing_twox_64_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-
+	instanceContext := env.(*runtime.Context)
+	dataSpan := args[0].I64()
 	data := asMemorySlice(instanceContext, dataSpan)
 
 	hash, err := common.Twox64(data)
 	if err != nil {
 		logger.Errorf("failed hashing data: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
 	logger.Debugf(
@@ -1480,75 +1478,80 @@ func ext_hashing_twox_64_version_1(context unsafe.Pointer, dataSpan C.int64_t) C
 	out, err := toWasmMemorySized(instanceContext, hash)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
-	return C.int32_t(out)
+	return []wasmer.Value{wasmer.NewI32(int32(out))}, nil
 }
 
 //export ext_offchain_index_set_version_1
-func ext_offchain_index_set_version_1(context unsafe.Pointer, keySpan, valueSpan C.int64_t) {
+func ext_offchain_index_set_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
+	instanceContext := env.(*runtime.Context)
+	keySpan := args[0].I64()
+	valueSpan := args[1].I64()
 
 	storageKey := asMemorySlice(instanceContext, keySpan)
 	newValue := asMemorySlice(instanceContext, valueSpan)
 	cp := make([]byte, len(newValue))
 	copy(cp, newValue)
 
-	err := runtimeCtx.NodeStorage.BaseDB.Put(storageKey, cp)
+	err := instanceContext.NodeStorage.BaseDB.Put(storageKey, cp)
 	if err != nil {
 		logger.Errorf("failed to set value in raw storage: %s", err)
 	}
+	return nil, nil
 }
 
 //export ext_offchain_local_storage_clear_version_1
-func ext_offchain_local_storage_clear_version_1(context unsafe.Pointer, kind C.int32_t, key C.int64_t) {
+func ext_offchain_local_storage_clear_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
+	instanceContext := env.(*runtime.Context)
+	kind := args[0].I32()
+	key := args[1].I64()
 
 	storageKey := asMemorySlice(instanceContext, key)
 
-	memory := instanceContext.Memory().Data()
+	memory := instanceContext.Memory.Data()
 	kindInt := binary.LittleEndian.Uint32(memory[kind : kind+4])
 
 	var err error
 
 	switch runtime.NodeStorageType(kindInt) {
 	case runtime.NodeStorageTypePersistent:
-		err = runtimeCtx.NodeStorage.PersistentStorage.Del(storageKey)
+		err = instanceContext.NodeStorage.PersistentStorage.Del(storageKey)
 	case runtime.NodeStorageTypeLocal:
-		err = runtimeCtx.NodeStorage.LocalStorage.Del(storageKey)
+		err = instanceContext.NodeStorage.LocalStorage.Del(storageKey)
 	}
 
 	if err != nil {
 		logger.Errorf("failed to clear value from storage: %s", err)
 	}
+	return nil, nil
 }
 
 //export ext_offchain_is_validator_version_1
-func ext_offchain_is_validator_version_1(context unsafe.Pointer) C.int32_t {
+func ext_offchain_is_validator_version_1(env interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	if runtimeCtx.Validator {
-		return 1
+	instanceContext := env.(*runtime.Context)
+	if instanceContext.Validator {
+		return []wasmer.Value{wasmer.NewI32(int32(1))}, nil
 	}
-	return 0
+	return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 }
 
 //export ext_offchain_local_storage_compare_and_set_version_1
-func ext_offchain_local_storage_compare_and_set_version_1(context unsafe.Pointer,
-	kind C.int32_t, key, oldValue, newValue C.int64_t) (newValueSet C.int32_t) {
+func ext_offchain_local_storage_compare_and_set_version_1(env interface{},
+	args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
+	runtimeCtx := env.(*runtime.Context)
+	kind := args[0].I32()
+	key := args[1].I64()
+	oldValue := args[2].I64()
+	newValue := args[3].I64()
 
-	storageKey := asMemorySlice(instanceContext, key)
+	storageKey := asMemorySlice(runtimeCtx, key)
 
 	var storedValue []byte
 	var err error
@@ -1562,31 +1565,32 @@ func ext_offchain_local_storage_compare_and_set_version_1(context unsafe.Pointer
 
 	if err != nil {
 		logger.Errorf("failed to get value from storage: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 	}
 
-	oldVal := asMemorySlice(instanceContext, oldValue)
-	newVal := asMemorySlice(instanceContext, newValue)
+	oldVal := asMemorySlice(runtimeCtx, oldValue)
+	newVal := asMemorySlice(runtimeCtx, newValue)
 	if reflect.DeepEqual(storedValue, oldVal) {
 		cp := make([]byte, len(newVal))
 		copy(cp, newVal)
 		err = runtimeCtx.NodeStorage.LocalStorage.Put(storageKey, cp)
 		if err != nil {
 			logger.Errorf("failed to set value in storage: %s", err)
-			return 0
+			return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 		}
 	}
 
-	return 1
+	return []wasmer.Value{wasmer.NewI32(int32(1))}, nil
 }
 
 //export ext_offchain_local_storage_get_version_1
-func ext_offchain_local_storage_get_version_1(context unsafe.Pointer, kind C.int32_t, key C.int64_t) C.int64_t {
+func ext_offchain_local_storage_get_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	storageKey := asMemorySlice(instanceContext, key)
+	runtimeCtx := env.(*runtime.Context)
+	kind := args[0].I32()
+	key := args[1].I64()
+	storageKey := asMemorySlice(runtimeCtx, key)
 
 	var res []byte
 	var err error
@@ -1602,22 +1606,25 @@ func ext_offchain_local_storage_get_version_1(context unsafe.Pointer, kind C.int
 		logger.Errorf("failed to get value from storage: %s", err)
 	}
 	// allocate memory for value and copy value to memory
-	ptr, err := toWasmMemoryOptional(instanceContext, res)
+	ptr, err := toWasmMemoryOptional(runtimeCtx, res)
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
-	return C.int64_t(ptr)
+	return []wasmer.Value{wasmer.NewI64(int32(ptr))}, nil
 }
 
 //export ext_offchain_local_storage_set_version_1
-func ext_offchain_local_storage_set_version_1(context unsafe.Pointer, kind C.int32_t, key, value C.int64_t) {
+func ext_offchain_local_storage_set_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	storageKey := asMemorySlice(instanceContext, key)
-	newValue := asMemorySlice(instanceContext, value)
+	runtimeCtx := env.(*runtime.Context)
+	kind := args[0].I32()
+	key := args[1].I64()
+	value := args[2].I64()
+
+	storageKey := asMemorySlice(runtimeCtx, key)
+	newValue := asMemorySlice(runtimeCtx, value)
 	cp := make([]byte, len(newValue))
 	copy(cp, newValue)
 
@@ -1632,37 +1639,37 @@ func ext_offchain_local_storage_set_version_1(context unsafe.Pointer, kind C.int
 	if err != nil {
 		logger.Errorf("failed to set value in storage: %s", err)
 	}
+	return nil, nil
 }
 
 //export ext_offchain_network_state_version_1
-func ext_offchain_network_state_version_1(context unsafe.Pointer) C.int64_t {
+func ext_offchain_network_state_version_1(env interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
+	runtimeCtx := env.(*runtime.Context)
 	if runtimeCtx.Network == nil {
-		return 0
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
 
 	nsEnc, err := scale.Marshal(runtimeCtx.Network.NetworkState())
 	if err != nil {
 		logger.Errorf("failed at encoding network state: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
 
 	// allocate memory for value and copy value to memory
-	ptr, err := toWasmMemorySized(instanceContext, nsEnc)
+	ptr, err := toWasmMemorySized(runtimeCtx, nsEnc)
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
 
-	return C.int64_t(ptr)
+	return []wasmer.Value{wasmer.NewI64(int64(ptr))}, nil
 }
 
 //export ext_offchain_random_seed_version_1
-func ext_offchain_random_seed_version_1(context unsafe.Pointer) C.int32_t {
+func ext_offchain_random_seed_version_1(env interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
+	instanceContext := env.(*runtime.Context)
 
 	seed := make([]byte, 32)
 	_, err := rand.Read(seed)
@@ -1673,14 +1680,15 @@ func ext_offchain_random_seed_version_1(context unsafe.Pointer) C.int32_t {
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
 	}
-	return C.int32_t(ptr)
+	return []wasmer.Value{wasmer.NewI32(ptr)}, nil
 }
 
 //export ext_offchain_submit_transaction_version_1
-func ext_offchain_submit_transaction_version_1(context unsafe.Pointer, data C.int64_t) C.int64_t {
+func ext_offchain_submit_transaction_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
+	instanceContext := env.(*runtime.Context)
+	data := args[0].I64()
 	extBytes := asMemorySlice(instanceContext, data)
 
 	var extrinsic []byte
@@ -1693,48 +1701,49 @@ func ext_offchain_submit_transaction_version_1(context unsafe.Pointer, data C.in
 	txv := transaction.NewValidity(0, [][]byte{{}}, [][]byte{{}}, 0, false)
 	vtx := transaction.NewValidTransaction(extrinsic, txv)
 
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	runtimeCtx.Transaction.AddToPool(vtx)
+	instanceContext.Transaction.AddToPool(vtx)
 
 	ptr, err := toWasmMemoryOptionalNil(instanceContext)
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
 	}
-	return ptr
+	return ptr, nil
 }
 
 //export ext_offchain_timestamp_version_1
-func ext_offchain_timestamp_version_1(_ unsafe.Pointer) C.int64_t {
+func ext_offchain_timestamp_version_1(_ interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
 	now := time.Now().Unix()
-	return C.int64_t(now)
+	return []wasmer.Value{wasmer.NewI64(now)}, nil
 }
 
 //export ext_offchain_sleep_until_version_1
-func ext_offchain_sleep_until_version_1(_ unsafe.Pointer, deadline C.int64_t) {
+func ext_offchain_sleep_until_version_1(_ interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-
-	dur := time.Until(time.UnixMilli(int64(deadline)))
+	deadline := args[0].I64()
+	dur := time.Until(time.UnixMilli(deadline))
 	if dur > 0 {
 		time.Sleep(dur)
 	}
+	return nil, nil
 }
 
 //export ext_offchain_http_request_start_version_1
-func ext_offchain_http_request_start_version_1(context unsafe.Pointer,
-	methodSpan, uriSpan, metaSpan C.int64_t) (pointerSize C.int64_t) {
+func ext_offchain_http_request_start_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
+	instanceContext := env.(*runtime.Context)
+	methodSpan := args[0].I64()
+	uriSpan := args[1].I64()
+	_ = args[2].I64() // metaSpan - unused
 
 	httpMethod := asMemorySlice(instanceContext, methodSpan)
 	uri := asMemorySlice(instanceContext, uriSpan)
 
 	result := scale.NewResult(int16(0), nil)
 
-	reqID, err := runtimeCtx.OffchainHTTPSet.StartRequest(string(httpMethod), string(uri))
+	reqID, err := instanceContext.OffchainHTTPSet.StartRequest(string(httpMethod), string(uri))
 	if err != nil {
 		// StartRequest error already was logged
 		logger.Errorf("failed to start request: %s", err)
@@ -1746,35 +1755,36 @@ func ext_offchain_http_request_start_version_1(context unsafe.Pointer,
 	// note: just check if an error occurs while setting the result data
 	if err != nil {
 		logger.Errorf("failed to set the result data: %s", err)
-		return C.int64_t(0)
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
 
 	enc, err := scale.Marshal(result)
 	if err != nil {
 		logger.Errorf("failed to scale marshal the result: %s", err)
-		return C.int64_t(0)
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
 
 	ptr, err := toWasmMemory(instanceContext, enc)
 	if err != nil {
 		logger.Errorf("failed to allocate result on memory: %s", err)
-		return C.int64_t(0)
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
 
-	return C.int64_t(ptr)
+	return []wasmer.Value{wasmer.NewI64(ptr)}, nil
 }
 
 //export ext_offchain_http_request_add_header_version_1
-func ext_offchain_http_request_add_header_version_1(context unsafe.Pointer,
-	reqID C.int32_t, nameSpan, valueSpan C.int64_t) (pointerSize C.int64_t) {
+func ext_offchain_http_request_add_header_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
+	instanceContext := env.(*runtime.Context)
+	reqID := args[0].I32()
+	nameSpan := args[1].I64()
+	valueSpan := args[2].I64()
 
 	name := asMemorySlice(instanceContext, nameSpan)
 	value := asMemorySlice(instanceContext, valueSpan)
 
-	runtimeCtx := instanceContext.Data().(*runtime.Context)
-	offchainReq := runtimeCtx.OffchainHTTPSet.Get(int16(reqID))
+	offchainReq := instanceContext.OffchainHTTPSet.Get(int16(reqID))
 
 	result := scale.NewResult(nil, nil)
 	resultMode := scale.OK
@@ -1788,30 +1798,31 @@ func ext_offchain_http_request_add_header_version_1(context unsafe.Pointer,
 	err = result.Set(resultMode, nil)
 	if err != nil {
 		logger.Errorf("failed to set the result data: %s", err)
-		return C.int64_t(0)
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
 
 	enc, err := scale.Marshal(result)
 	if err != nil {
 		logger.Errorf("failed to scale marshal the result: %s", err)
-		return C.int64_t(0)
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
 
 	ptr, err := toWasmMemory(instanceContext, enc)
 	if err != nil {
 		logger.Errorf("failed to allocate result on memory: %s", err)
-		return C.int64_t(0)
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
 
-	return C.int64_t(ptr)
+	return []wasmer.Value{wasmer.NewI64(ptr)}, nil
 }
 
 //export ext_storage_append_version_1
-func ext_storage_append_version_1(context unsafe.Pointer, keySpan, valueSpan C.int64_t) {
+func ext_storage_append_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	keySpan := args[0].I64()
+	valueSpan := args[1].I64()
+	storage := instanceContext.Storage
 
 	key := asMemorySlice(instanceContext, keySpan)
 	valueAppend := asMemorySlice(instanceContext, valueSpan)
@@ -1826,59 +1837,64 @@ func ext_storage_append_version_1(context unsafe.Pointer, keySpan, valueSpan C.i
 	if err != nil {
 		logger.Errorf("failed appending to storage: %s", err)
 	}
+	return nil, nil
 }
 
 //export ext_storage_changes_root_version_1
-func ext_storage_changes_root_version_1(context unsafe.Pointer, parentHashSpan C.int64_t) C.int64_t {
+func ext_storage_changes_root_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 	logger.Debug("returning None")
 
-	instanceContext := wasm.IntoInstanceContext(context)
+	instanceContext := env.(*runtime.Context)
+	_ = args[0].I64() // parentHashSpan - unused
 
 	rootSpan, err := toWasmMemoryOptionalNil(instanceContext)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(int64(0))}, nil
 	}
 
-	return rootSpan
+	return rootSpan, nil
 }
 
 //export ext_storage_clear_version_1
-func ext_storage_clear_version_1(context unsafe.Pointer, keySpan C.int64_t) {
+func ext_storage_clear_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	keySpan := args[0].I64()
+	storage := instanceContext.Storage
 
 	key := asMemorySlice(instanceContext, keySpan)
 
 	logger.Debugf("key: 0x%x", key)
 	err := storage.Delete(key)
 	panicOnError(err)
+	return nil, nil
 }
 
 //export ext_storage_clear_prefix_version_1
-func ext_storage_clear_prefix_version_1(context unsafe.Pointer, prefixSpan C.int64_t) {
+func ext_storage_clear_prefix_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	prefixSpan := args[0].I64()
+	storage := instanceContext.Storage
 
 	prefix := asMemorySlice(instanceContext, prefixSpan)
 	logger.Debugf("prefix: 0x%x", prefix)
 
 	err := storage.ClearPrefix(prefix)
 	panicOnError(err)
+	return nil, nil
 }
 
 //export ext_storage_clear_prefix_version_2
-func ext_storage_clear_prefix_version_2(context unsafe.Pointer, prefixSpan, lim C.int64_t) C.int64_t {
+func ext_storage_clear_prefix_version_2(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	prefixSpan := args[0].I64()
+	lim := args[1].I64()
+	storage := instanceContext.Storage
 
 	prefix := asMemorySlice(instanceContext, prefixSpan)
 	logger.Debugf("prefix: 0x%x", prefix)
@@ -1889,7 +1905,7 @@ func ext_storage_clear_prefix_version_2(context unsafe.Pointer, prefixSpan, lim 
 	err := scale.Unmarshal(limitBytes, &limit)
 	if err != nil {
 		logger.Warnf("failed scale decoding limit: %s", err)
-		return mustToWasmMemoryNil(instanceContext)
+		return mustToWasmMemoryNil(instanceContext), nil
 	}
 
 	if len(limit) == 0 {
@@ -1901,47 +1917,49 @@ func ext_storage_clear_prefix_version_2(context unsafe.Pointer, prefixSpan, lim 
 	numRemoved, all, err := storage.ClearPrefixLimit(prefix, limitUint)
 	if err != nil {
 		logger.Errorf("failed to clear prefix limit: %s", err)
-		return mustToWasmMemoryNil(instanceContext)
+		return mustToWasmMemoryNil(instanceContext), nil
 	}
 
 	encBytes, err := toKillStorageResultEnum(all, numRemoved)
 	if err != nil {
 		logger.Errorf("failed to allocate memory: %s", err)
-		return mustToWasmMemoryNil(instanceContext)
+		return mustToWasmMemoryNil(instanceContext), nil
 	}
 
 	valueSpan, err := toWasmMemory(instanceContext, encBytes)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return mustToWasmMemoryNil(instanceContext)
+		return mustToWasmMemoryNil(instanceContext), nil
 	}
 
-	return C.int64_t(valueSpan)
+	return []wasmer.Value{wasmer.NewI64(valueSpan)}, nil
 }
 
 //export ext_storage_exists_version_1
-func ext_storage_exists_version_1(context unsafe.Pointer, keySpan C.int64_t) C.int32_t {
+func ext_storage_exists_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	storage := instanceContext.Data().(*runtime.Context).Storage
+	instanceContext := env.(*runtime.Context)
+	keySpan := args[0].I64()
+	storage := instanceContext.Storage
 
 	key := asMemorySlice(instanceContext, keySpan)
 	logger.Debugf("key: 0x%x", key)
 
 	value := storage.Get(key)
 	if value != nil {
-		return 1
+		return []wasmer.Value{wasmer.NewI32(int32(1))}, nil
 	}
 
-	return 0
+	return []wasmer.Value{wasmer.NewI32(int32(0))}, nil
 }
 
 //export ext_storage_get_version_1
-func ext_storage_get_version_1(context unsafe.Pointer, keySpan C.int64_t) C.int64_t {
+func ext_storage_get_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	storage := instanceContext.Data().(*runtime.Context).Storage
+	instanceContext := env.(*runtime.Context)
+	keySpan := args[0].I64()
+	storage := instanceContext.Storage
 
 	key := asMemorySlice(instanceContext, keySpan)
 	logger.Debugf("key: 0x%x", key)
@@ -1952,18 +1970,19 @@ func ext_storage_get_version_1(context unsafe.Pointer, keySpan C.int64_t) C.int6
 	valueSpan, err := toWasmMemoryOptional(instanceContext, value)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return mustToWasmMemoryOptionalNil(instanceContext)
+		return mustToWasmMemoryOptionalNil(instanceContext), nil
 	}
 
-	return C.int64_t(valueSpan)
+	return []wasmer.Value{wasmer.NewI64(valueSpan)}, nil
 }
 
 //export ext_storage_next_key_version_1
-func ext_storage_next_key_version_1(context unsafe.Pointer, keySpan C.int64_t) C.int64_t {
+func ext_storage_next_key_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	storage := instanceContext.Data().(*runtime.Context).Storage
+	instanceContext := env.(*runtime.Context)
+	keySpan := args[0].I64()
+	storage := instanceContext.Storage
 
 	key := asMemorySlice(instanceContext, keySpan)
 
@@ -1975,19 +1994,22 @@ func ext_storage_next_key_version_1(context unsafe.Pointer, keySpan C.int64_t) C
 	nextSpan, err := toWasmMemoryOptional(instanceContext, next)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(nextSpan)
+	return []wasmer.Value{wasmer.NewI64(nextSpan)}, nil
 }
 
 //export ext_storage_read_version_1
-func ext_storage_read_version_1(context unsafe.Pointer, keySpan, valueOut C.int64_t, offset C.int32_t) C.int64_t {
+func ext_storage_read_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	storage := instanceContext.Data().(*runtime.Context).Storage
-	memory := instanceContext.Memory().Data()
+	instanceContext := env.(*runtime.Context)
+	keySpan := args[0].I64()
+	valueOut := args[1].I64()
+	offset := args[2].I32()
+	storage := instanceContext.Storage
+	memory := instanceContext.Memory.Data()
 
 	key := asMemorySlice(instanceContext, keySpan)
 	value := storage.Get(key)
@@ -1996,7 +2018,8 @@ func ext_storage_read_version_1(context unsafe.Pointer, keySpan, valueOut C.int6
 		key, value)
 
 	if value == nil {
-		return mustToWasmMemoryOptionalNil(instanceContext)
+
+		return mustToWasmMemoryOptionalNil(instanceContext), nil
 	}
 
 	var size uint32
@@ -2009,23 +2032,23 @@ func ext_storage_read_version_1(context unsafe.Pointer, keySpan, valueOut C.int6
 	sizeSpan, err := toWasmMemoryOptionalUint32(instanceContext, &size)
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(sizeSpan)
+	return []wasmer.Value{wasmer.NewI64(sizeSpan)}, nil
 }
 
 //export ext_storage_root_version_1
-func ext_storage_root_version_1(context unsafe.Pointer) C.int64_t {
+func ext_storage_root_version_1(env interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	storage := instanceContext.Data().(*runtime.Context).Storage
+	instanceContext := env.(*runtime.Context)
+	storage := instanceContext.Storage
 
 	root, err := storage.Root()
 	if err != nil {
 		logger.Errorf("failed to get storage root: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
 	logger.Debugf("root hash is: %s", root)
@@ -2033,25 +2056,28 @@ func ext_storage_root_version_1(context unsafe.Pointer) C.int64_t {
 	rootSpan, err := toWasmMemory(instanceContext, root[:])
 	if err != nil {
 		logger.Errorf("failed to allocate: %s", err)
-		return 0
+		return []wasmer.Value{wasmer.NewI64(0)}, nil
 	}
 
-	return C.int64_t(rootSpan)
+	return []wasmer.Value{wasmer.NewI64(rootSpan)}, nil
 }
 
 //export ext_storage_root_version_2
-func ext_storage_root_version_2(context unsafe.Pointer, version C.int32_t) C.int64_t {
+func ext_storage_root_version_2(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	// TODO: update to use state trie version 1 (#2418)
-	return ext_storage_root_version_1(context)
+	instanceContext := env.(*runtime.Context)
+	_ = args[0].I32() // version - unused
+	return ext_storage_root_version_1(instanceContext, args)
 }
 
 //export ext_storage_set_version_1
-func ext_storage_set_version_1(context unsafe.Pointer, keySpan, valueSpan C.int64_t) {
+func ext_storage_set_version_1(env interface{}, args []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Trace("executing...")
 
-	instanceContext := wasm.IntoInstanceContext(context)
-	ctx := instanceContext.Data().(*runtime.Context)
-	storage := ctx.Storage
+	instanceContext := env.(*runtime.Context)
+	keySpan := args[0].I64()
+	valueSpan := args[1].I64()
+	storage := instanceContext.Storage
 
 	key := asMemorySlice(instanceContext, keySpan)
 	value := asMemorySlice(instanceContext, valueSpan)
@@ -2064,128 +2090,29 @@ func ext_storage_set_version_1(context unsafe.Pointer, keySpan, valueSpan C.int6
 		key, value)
 	err := storage.Put(key, cp)
 	panicOnError(err)
+	return nil, nil
 }
 
 //export ext_storage_start_transaction_version_1
-func ext_storage_start_transaction_version_1(context unsafe.Pointer) {
+func ext_storage_start_transaction_version_1(env interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	instanceContext.Data().(*runtime.Context).Storage.BeginStorageTransaction()
+	instanceContext := env.(*runtime.Context)
+	instanceContext.Storage.BeginStorageTransaction()
+	return nil, nil
 }
 
 //export ext_storage_rollback_transaction_version_1
-func ext_storage_rollback_transaction_version_1(context unsafe.Pointer) {
+func ext_storage_rollback_transaction_version_1(env interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	instanceContext.Data().(*runtime.Context).Storage.RollbackStorageTransaction()
+	instanceContext := env.(*runtime.Context)
+	instanceContext.Storage.RollbackStorageTransaction()
+	return nil, nil
 }
 
 //export ext_storage_commit_transaction_version_1
-func ext_storage_commit_transaction_version_1(context unsafe.Pointer) {
+func ext_storage_commit_transaction_version_1(env interface{}, _ []wasmer.Value) ([]wasmer.Value, error) {
 	logger.Debug("executing...")
-	instanceContext := wasm.IntoInstanceContext(context)
-	instanceContext.Data().(*runtime.Context).Storage.CommitStorageTransaction()
-}
-
-// importsNodeRuntime returns the WASM imports for the node runtime.
-func importsNodeRuntime() (imports *wasm.Imports, err error) {
-	imports = wasm.NewImports()
-	// Note imports are closed by the call to wasm.Instance.Close()
-
-	for _, toRegister := range []struct {
-		importName     string
-		implementation interface{}
-		cgoPointer     unsafe.Pointer
-	}{
-		{"ext_allocator_free_version_1", ext_allocator_free_version_1, C.ext_allocator_free_version_1},
-		{"ext_allocator_malloc_version_1", ext_allocator_malloc_version_1, C.ext_allocator_malloc_version_1},
-		{"ext_crypto_ecdsa_verify_version_2", ext_crypto_ecdsa_verify_version_2, C.ext_crypto_ecdsa_verify_version_2},
-		{"ext_crypto_ed25519_generate_version_1", ext_crypto_ed25519_generate_version_1, C.ext_crypto_ed25519_generate_version_1},
-		{"ext_crypto_ed25519_public_keys_version_1", ext_crypto_ed25519_public_keys_version_1, C.ext_crypto_ed25519_public_keys_version_1},
-		{"ext_crypto_ed25519_sign_version_1", ext_crypto_ed25519_sign_version_1, C.ext_crypto_ed25519_sign_version_1},
-		{"ext_crypto_ed25519_verify_version_1", ext_crypto_ed25519_verify_version_1, C.ext_crypto_ed25519_verify_version_1},
-		{"ext_crypto_finish_batch_verify_version_1", ext_crypto_finish_batch_verify_version_1, C.ext_crypto_finish_batch_verify_version_1},
-		{"ext_crypto_secp256k1_ecdsa_recover_compressed_version_1", ext_crypto_secp256k1_ecdsa_recover_compressed_version_1, C.ext_crypto_secp256k1_ecdsa_recover_compressed_version_1},
-		{"ext_crypto_secp256k1_ecdsa_recover_compressed_version_2", ext_crypto_secp256k1_ecdsa_recover_compressed_version_2, C.ext_crypto_secp256k1_ecdsa_recover_compressed_version_2},
-		{"ext_crypto_secp256k1_ecdsa_recover_version_1", ext_crypto_secp256k1_ecdsa_recover_version_1, C.ext_crypto_secp256k1_ecdsa_recover_version_1},
-		{"ext_crypto_secp256k1_ecdsa_recover_version_2", ext_crypto_secp256k1_ecdsa_recover_version_2, C.ext_crypto_secp256k1_ecdsa_recover_version_2},
-		{"ext_crypto_sr25519_generate_version_1", ext_crypto_sr25519_generate_version_1, C.ext_crypto_sr25519_generate_version_1},
-		{"ext_crypto_sr25519_public_keys_version_1", ext_crypto_sr25519_public_keys_version_1, C.ext_crypto_sr25519_public_keys_version_1},
-		{"ext_crypto_sr25519_sign_version_1", ext_crypto_sr25519_sign_version_1, C.ext_crypto_sr25519_sign_version_1},
-		{"ext_crypto_sr25519_verify_version_1", ext_crypto_sr25519_verify_version_1, C.ext_crypto_sr25519_verify_version_1},
-		{"ext_crypto_sr25519_verify_version_2", ext_crypto_sr25519_verify_version_2, C.ext_crypto_sr25519_verify_version_2},
-		{"ext_crypto_start_batch_verify_version_1", ext_crypto_start_batch_verify_version_1, C.ext_crypto_start_batch_verify_version_1},
-		{"ext_default_child_storage_clear_prefix_version_1", ext_default_child_storage_clear_prefix_version_1, C.ext_default_child_storage_clear_prefix_version_1},
-		{"ext_default_child_storage_clear_version_1", ext_default_child_storage_clear_version_1, C.ext_default_child_storage_clear_version_1},
-		{"ext_default_child_storage_exists_version_1", ext_default_child_storage_exists_version_1, C.ext_default_child_storage_exists_version_1},
-		{"ext_default_child_storage_get_version_1", ext_default_child_storage_get_version_1, C.ext_default_child_storage_get_version_1},
-		{"ext_default_child_storage_next_key_version_1", ext_default_child_storage_next_key_version_1, C.ext_default_child_storage_next_key_version_1},
-		{"ext_default_child_storage_read_version_1", ext_default_child_storage_read_version_1, C.ext_default_child_storage_read_version_1},
-		{"ext_default_child_storage_root_version_1", ext_default_child_storage_root_version_1, C.ext_default_child_storage_root_version_1},
-		{"ext_default_child_storage_set_version_1", ext_default_child_storage_set_version_1, C.ext_default_child_storage_set_version_1},
-		{"ext_default_child_storage_storage_kill_version_1", ext_default_child_storage_storage_kill_version_1, C.ext_default_child_storage_storage_kill_version_1},
-		{"ext_default_child_storage_storage_kill_version_2", ext_default_child_storage_storage_kill_version_2, C.ext_default_child_storage_storage_kill_version_2},
-		{"ext_default_child_storage_storage_kill_version_3", ext_default_child_storage_storage_kill_version_3, C.ext_default_child_storage_storage_kill_version_3},
-		{"ext_hashing_blake2_128_version_1", ext_hashing_blake2_128_version_1, C.ext_hashing_blake2_128_version_1},
-		{"ext_hashing_blake2_256_version_1", ext_hashing_blake2_256_version_1, C.ext_hashing_blake2_256_version_1},
-		{"ext_hashing_keccak_256_version_1", ext_hashing_keccak_256_version_1, C.ext_hashing_keccak_256_version_1},
-		{"ext_hashing_sha2_256_version_1", ext_hashing_sha2_256_version_1, C.ext_hashing_sha2_256_version_1},
-		{"ext_hashing_twox_128_version_1", ext_hashing_twox_128_version_1, C.ext_hashing_twox_128_version_1},
-		{"ext_hashing_twox_256_version_1", ext_hashing_twox_256_version_1, C.ext_hashing_twox_256_version_1},
-		{"ext_hashing_twox_64_version_1", ext_hashing_twox_64_version_1, C.ext_hashing_twox_64_version_1},
-		{"ext_logging_log_version_1", ext_logging_log_version_1, C.ext_logging_log_version_1},
-		{"ext_logging_max_level_version_1", ext_logging_max_level_version_1, C.ext_logging_max_level_version_1},
-		{"ext_misc_print_hex_version_1", ext_misc_print_hex_version_1, C.ext_misc_print_hex_version_1},
-		{"ext_misc_print_num_version_1", ext_misc_print_num_version_1, C.ext_misc_print_num_version_1},
-		{"ext_misc_print_utf8_version_1", ext_misc_print_utf8_version_1, C.ext_misc_print_utf8_version_1},
-		{"ext_misc_runtime_version_version_1", ext_misc_runtime_version_version_1, C.ext_misc_runtime_version_version_1},
-		{"ext_offchain_http_request_add_header_version_1", ext_offchain_http_request_add_header_version_1, C.ext_offchain_http_request_add_header_version_1},
-		{"ext_offchain_http_request_start_version_1", ext_offchain_http_request_start_version_1, C.ext_offchain_http_request_start_version_1},
-		{"ext_offchain_index_set_version_1", ext_offchain_index_set_version_1, C.ext_offchain_index_set_version_1},
-		{"ext_offchain_is_validator_version_1", ext_offchain_is_validator_version_1, C.ext_offchain_is_validator_version_1},
-		{"ext_offchain_local_storage_clear_version_1", ext_offchain_local_storage_clear_version_1, C.ext_offchain_local_storage_clear_version_1},
-		{"ext_offchain_local_storage_compare_and_set_version_1", ext_offchain_local_storage_compare_and_set_version_1, C.ext_offchain_local_storage_compare_and_set_version_1},
-		{"ext_offchain_local_storage_get_version_1", ext_offchain_local_storage_get_version_1, C.ext_offchain_local_storage_get_version_1},
-		{"ext_offchain_local_storage_set_version_1", ext_offchain_local_storage_set_version_1, C.ext_offchain_local_storage_set_version_1},
-		{"ext_offchain_network_state_version_1", ext_offchain_network_state_version_1, C.ext_offchain_network_state_version_1},
-		{"ext_offchain_random_seed_version_1", ext_offchain_random_seed_version_1, C.ext_offchain_random_seed_version_1},
-		{"ext_offchain_sleep_until_version_1", ext_offchain_sleep_until_version_1, C.ext_offchain_sleep_until_version_1},
-		{"ext_offchain_submit_transaction_version_1", ext_offchain_submit_transaction_version_1, C.ext_offchain_submit_transaction_version_1},
-		{"ext_offchain_timestamp_version_1", ext_offchain_timestamp_version_1, C.ext_offchain_timestamp_version_1},
-		{"ext_sandbox_instance_teardown_version_1", ext_sandbox_instance_teardown_version_1, C.ext_sandbox_instance_teardown_version_1},
-		{"ext_sandbox_instantiate_version_1", ext_sandbox_instantiate_version_1, C.ext_sandbox_instantiate_version_1},
-		{"ext_sandbox_invoke_version_1", ext_sandbox_invoke_version_1, C.ext_sandbox_invoke_version_1},
-		{"ext_sandbox_memory_get_version_1", ext_sandbox_memory_get_version_1, C.ext_sandbox_memory_get_version_1},
-		{"ext_sandbox_memory_new_version_1", ext_sandbox_memory_new_version_1, C.ext_sandbox_memory_new_version_1},
-		{"ext_sandbox_memory_set_version_1", ext_sandbox_memory_set_version_1, C.ext_sandbox_memory_set_version_1},
-		{"ext_sandbox_memory_teardown_version_1", ext_sandbox_memory_teardown_version_1, C.ext_sandbox_memory_teardown_version_1},
-		{"ext_storage_append_version_1", ext_storage_append_version_1, C.ext_storage_append_version_1},
-		{"ext_storage_changes_root_version_1", ext_storage_changes_root_version_1, C.ext_storage_changes_root_version_1},
-		{"ext_storage_clear_prefix_version_1", ext_storage_clear_prefix_version_1, C.ext_storage_clear_prefix_version_1},
-		{"ext_storage_clear_prefix_version_2", ext_storage_clear_prefix_version_2, C.ext_storage_clear_prefix_version_2},
-		{"ext_storage_clear_version_1", ext_storage_clear_version_1, C.ext_storage_clear_version_1},
-		{"ext_storage_commit_transaction_version_1", ext_storage_commit_transaction_version_1, C.ext_storage_commit_transaction_version_1},
-		{"ext_storage_exists_version_1", ext_storage_exists_version_1, C.ext_storage_exists_version_1},
-		{"ext_storage_get_version_1", ext_storage_get_version_1, C.ext_storage_get_version_1},
-		{"ext_storage_next_key_version_1", ext_storage_next_key_version_1, C.ext_storage_next_key_version_1},
-		{"ext_storage_read_version_1", ext_storage_read_version_1, C.ext_storage_read_version_1},
-		{"ext_storage_rollback_transaction_version_1", ext_storage_rollback_transaction_version_1, C.ext_storage_rollback_transaction_version_1},
-		{"ext_storage_root_version_1", ext_storage_root_version_1, C.ext_storage_root_version_1},
-		{"ext_storage_root_version_2", ext_storage_root_version_2, C.ext_storage_root_version_2},
-		{"ext_storage_set_version_1", ext_storage_set_version_1, C.ext_storage_set_version_1},
-		{"ext_storage_start_transaction_version_1", ext_storage_start_transaction_version_1, C.ext_storage_start_transaction_version_1},
-		{"ext_transaction_index_index_version_1", ext_transaction_index_index_version_1, C.ext_transaction_index_index_version_1},
-		{"ext_transaction_index_renew_version_1", ext_transaction_index_renew_version_1, C.ext_transaction_index_renew_version_1},
-		{"ext_trie_blake2_256_ordered_root_version_1", ext_trie_blake2_256_ordered_root_version_1, C.ext_trie_blake2_256_ordered_root_version_1},
-		{"ext_trie_blake2_256_ordered_root_version_2", ext_trie_blake2_256_ordered_root_version_2, C.ext_trie_blake2_256_ordered_root_version_2},
-		{"ext_trie_blake2_256_root_version_1", ext_trie_blake2_256_root_version_1, C.ext_trie_blake2_256_root_version_1},
-		{"ext_trie_blake2_256_verify_proof_version_1", ext_trie_blake2_256_verify_proof_version_1, C.ext_trie_blake2_256_verify_proof_version_1},
-	} {
-		_, err = imports.AppendFunction(toRegister.importName, toRegister.implementation, toRegister.cgoPointer)
-		if err != nil {
-			return nil, fmt.Errorf("importing function: %w", err)
-		}
-	}
-
-	return imports, nil
+	instanceContext := env.(*runtime.Context)
+	instanceContext.Storage.CommitStorageTransaction()
+	return nil, nil
 }
