+++ lib/runtime/wasmer/memory.go
@@ -1,7 +1,7 @@
 // Copyright 2023 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
 
-package newWasmer
+package wasmer
 
 import (
 	"errors"
